Network.AWS.MachineLearning

Description

Amazon Machine Learning (Amazon ML) is a robust machine learning platform that allows software developers to build and train predictive applications and host those applications in a scalable AWS cloud solution.

Documentation

module Network.AWS.MachineLearning.CreateBatchPrediction

module Network.AWS.MachineLearning.CreateDataSourceFromRDS

module Network.AWS.MachineLearning.CreateDataSourceFromRedshift

module Network.AWS.MachineLearning.CreateDataSourceFromS

module Network.AWS.MachineLearning.CreateEvaluation

module Network.AWS.MachineLearning.CreateMLModel

module Network.AWS.MachineLearning.CreateRealtimeEndpoint

module Network.AWS.MachineLearning.DeleteBatchPrediction

module Network.AWS.MachineLearning.DeleteDataSource

module Network.AWS.MachineLearning.DeleteEvaluation

module Network.AWS.MachineLearning.DeleteMLModel

module Network.AWS.MachineLearning.DeleteRealtimeEndpoint

module Network.AWS.MachineLearning.DescribeBatchPredictions

module Network.AWS.MachineLearning.DescribeDataSources

module Network.AWS.MachineLearning.DescribeEvaluations

module Network.AWS.MachineLearning.DescribeMLModels

module Network.AWS.MachineLearning.GetBatchPrediction

module Network.AWS.MachineLearning.GetDataSource

module Network.AWS.MachineLearning.GetEvaluation

module Network.AWS.MachineLearning.GetMLModel

module Network.AWS.MachineLearning.Predict

module Network.AWS.MachineLearning.Types

module Network.AWS.MachineLearning.UpdateBatchPrediction

module Network.AWS.MachineLearning.UpdateDataSource

module Network.AWS.MachineLearning.UpdateEvaluation

module Network.AWS.MachineLearning.UpdateMLModel