Network.AWS.MachineLearning.CreateBatchPrediction

Description

Generates predictions for a group of observations. The observations to process exist in one or more data files referenced by a DataSource. This operation creates a new BatchPrediction, and uses an MLModel and the data files referenced by the DataSource as information sources.

CreateBatchPrediction is an asynchronous operation. In response to CreateBatchPrediction, Amazon Machine Learning (Amazon ML) immediately returns and sets the BatchPrediction status to PENDING. After the BatchPrediction completes, Amazon ML sets the status to COMPLETED.

You can poll for status updates by using the GetBatchPrediction operation and checking the Status parameter of the result. After the COMPLETED status appears, the results are available in the location specified by the OutputUri parameter.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_CreateBatchPrediction.html

Request

data CreateBatchPrediction

Request constructor

createBatchPrediction

Arguments

:: Text
cbpBatchPredictionId

-> Text
cbpMLModelId

-> Text
cbpBatchPredictionDataSourceId

-> Text
cbpOutputUri

-> CreateBatchPrediction

CreateBatchPrediction constructor.

The fields accessible through corresponding lenses are:

cbpBatchPredictionDataSourceId :: Text

cbpBatchPredictionId :: Text

cbpBatchPredictionName :: Maybe Text

cbpMLModelId :: Text

cbpOutputUri :: Text

Request lenses

cbpBatchPredictionDataSourceId :: Lens' CreateBatchPrediction Text

The ID of the DataSource that points to the group of observations to predict.

cbpBatchPredictionId :: Lens' CreateBatchPrediction Text

A user-supplied ID that uniquely identifies the BatchPrediction.

cbpBatchPredictionName :: Lens' CreateBatchPrediction (Maybe Text)

A user-supplied name or description of the BatchPrediction. BatchPredictionName can only use the UTF-8 character set.

cbpMLModelId :: Lens' CreateBatchPrediction Text

The ID of the MLModel that will generate predictions for the group of observations.

cbpOutputUri :: Lens' CreateBatchPrediction Text

The location of an Amazon Simple Storage Service (Amazon S3) bucket or directory to store the batch prediction results. The following substrings are not allowed in the s3 key portion of the "outputURI" field: :, //, /./, /../.

Amazon ML needs permissions to store and retrieve the logs on your behalf. For information about how to set permissions, see the Amazon Machine LearningDeveloper Guide.

Response

data CreateBatchPredictionResponse

Response constructor

createBatchPredictionResponse :: CreateBatchPredictionResponse

CreateBatchPredictionResponse constructor.

The fields accessible through corresponding lenses are:

cbprBatchPredictionId :: Maybe Text

Response lenses

cbprBatchPredictionId :: Lens' CreateBatchPredictionResponse (Maybe Text)

A user-supplied ID that uniquely identifies the BatchPrediction. This value is identical to the value of the BatchPredictionId in the request.