Network.AWS.MachineLearning.CreateDataSourceFromRDS

Description

Creates a DataSource object from an Amazon Relational Database Service (Amazon RDS). A DataSource references data that can be used to perform CreateMLModel, CreateEvaluation, or CreateBatchPrediction operations.

CreateDataSourceFromRDS is an asynchronous operation. In response to CreateDataSourceFromRDS, Amazon Machine Learning (Amazon ML) immediately returns and sets the DataSource status to PENDING. After the DataSource is created and ready for use, Amazon ML sets the Status parameter to COMPLETED. DataSource in COMPLETED or PENDING status can only be used to perform CreateMLModel, CreateEvaluation, or CreateBatchPrediction operations.

If Amazon ML cannot accept the input source, it sets the Status parameter to FAILED and includes an error message in the Message attribute of the GetDataSource operation response.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_CreateDataSourceFromRDS.html

Request

data CreateDataSourceFromRDS

Request constructor

createDataSourceFromRDS

Arguments

:: Text
cdsfrdsDataSourceId

-> RDSDataSpec
cdsfrdsRDSData

-> Text
cdsfrdsRoleARN

-> CreateDataSourceFromRDS

CreateDataSourceFromRDS constructor.

The fields accessible through corresponding lenses are:

cdsfrdsComputeStatistics :: Maybe Bool

cdsfrdsDataSourceId :: Text

cdsfrdsDataSourceName :: Maybe Text

cdsfrdsRDSData :: RDSDataSpec

cdsfrdsRoleARN :: Text

Request lenses

cdsfrdsComputeStatistics :: Lens' CreateDataSourceFromRDS (Maybe Bool)

The compute statistics for a DataSource. The statistics are generated from the observation data referenced by a DataSource. Amazon ML uses the statistics internally during an MLModel training. This parameter must be set to true if the 'DataSource' needs to be used for MLModel training.

cdsfrdsDataSourceId :: Lens' CreateDataSourceFromRDS Text

A user-supplied ID that uniquely identifies the DataSource. Typically, an Amazon Resource Number (ARN) becomes the ID for a DataSource.

cdsfrdsDataSourceName :: Lens' CreateDataSourceFromRDS (Maybe Text)

A user-supplied name or description of the DataSource.

cdsfrdsRDSData :: Lens' CreateDataSourceFromRDS RDSDataSpec

The data specification of an Amazon RDS DataSource:

DatabaseInformation - 'DatabaseName ' - Name of the Amazon RDS database. 'InstanceIdentifier ' - Unique identifier for the Amazon RDS database instance.

DatabaseCredentials - AWS Identity and Access Management (IAM) credentials that are used to connect to the Amazon RDS database.

ResourceRole - Role (DataPipelineDefaultResourceRole) assumed by an Amazon Elastic Compute Cloud (EC2) instance to carry out the copy task from Amazon RDS to Amazon S3. For more information, see Role templates for data pipelines.

ServiceRole - Role (DataPipelineDefaultRole) assumed by the AWS Data Pipeline service to monitor the progress of the copy task from Amazon RDS to Amazon Simple Storage Service (S3). For more information, see Role templates for data pipelines.

SecurityInfo - Security information to use to access an Amazon RDS instance. You need to set up appropriate ingress rules for the security entity IDs provided to allow access to the Amazon RDS instance. Specify a [SubnetId, SecurityGroupIds] pair for a VPC-based Amazon RDS instance.

SelectSqlQuery - Query that is used to retrieve the observation data for the Datasource.

S3StagingLocation - Amazon S3 location for staging RDS data. The data retrieved from Amazon RDS using SelectSqlQuery is stored in this location.

DataSchemaUri - Amazon S3 location of the DataSchema.

DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified.

DataRearrangement - A JSON string representing the splitting requirement of a Datasource.

Sample - ' "{"randomSeed":"some-random-seed","splitting":{"percentBegin":10,"percentEnd":60}}"'

cdsfrdsRoleARN :: Lens' CreateDataSourceFromRDS Text

The role that Amazon ML assumes on behalf of the user to create and activate a data pipeline in the user’s account and copy data (using the SelectSqlQuery) query from Amazon RDS to Amazon S3.

Response

data CreateDataSourceFromRDSResponse

Response constructor

createDataSourceFromRDSResponse :: CreateDataSourceFromRDSResponse

CreateDataSourceFromRDSResponse constructor.

The fields accessible through corresponding lenses are:

cdsfrdsrDataSourceId :: Maybe Text

Response lenses

cdsfrdsrDataSourceId :: Lens' CreateDataSourceFromRDSResponse (Maybe Text)

A user-supplied ID that uniquely identifies the datasource. This value should be identical to the value of the DataSourceID in the request.