Network.AWS.MachineLearning.CreateEvaluation

Description

Creates a new Evaluation of an MLModel. An MLModel is evaluated on a set of observations associated to a DataSource. Like a DataSource for an MLModel, the DataSource for an Evaluation contains values for the Target Variable. The Evaluation compares the predicted result for each observation to the actual outcome and provides a summary so that you know how effective the MLModel functions on the test data. Evaluation generates a relevant performance metric such as BinaryAUC, RegressionRMSE or MulticlassAvgFScore based on the corresponding MLModelType: BINARY, REGRESSION or MULTICLASS.

CreateEvaluation is an asynchronous operation. In response to CreateEvaluation, Amazon Machine Learning (Amazon ML) immediately returns and sets the evaluation status to PENDING. After the Evaluation is created and ready for use, Amazon ML sets the status to COMPLETED.

You can use the GetEvaluation operation to check progress of the evaluation during the creation operation.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_CreateEvaluation.html

Request

data CreateEvaluation

Request constructor

createEvaluation

Arguments

:: Text
ceEvaluationId

-> Text
ceMLModelId

-> Text
ceEvaluationDataSourceId

-> CreateEvaluation

CreateEvaluation constructor.

The fields accessible through corresponding lenses are:

ceEvaluationDataSourceId :: Text

ceEvaluationId :: Text

ceEvaluationName :: Maybe Text

ceMLModelId :: Text

Request lenses

ceEvaluationDataSourceId :: Lens' CreateEvaluation Text

The ID of the DataSource for the evaluation. The schema of the DataSource must match the schema used to create the MLModel.

ceEvaluationId :: Lens' CreateEvaluation Text

A user-supplied ID that uniquely identifies the Evaluation.

ceEvaluationName :: Lens' CreateEvaluation (Maybe Text)

A user-supplied name or description of the Evaluation.

ceMLModelId :: Lens' CreateEvaluation Text

The ID of the MLModel to evaluate.

The schema used in creating the MLModel must match the schema of the DataSource used in the Evaluation.

Response

data CreateEvaluationResponse

Response constructor

createEvaluationResponse :: CreateEvaluationResponse

CreateEvaluationResponse constructor.

The fields accessible through corresponding lenses are:

cerEvaluationId :: Maybe Text

Response lenses

cerEvaluationId :: Lens' CreateEvaluationResponse (Maybe Text)

The user-supplied ID that uniquely identifies the Evaluation. This value should be identical to the value of the EvaluationId in the request.