Network.AWS.MachineLearning.DeleteBatchPrediction

Description

Assigns the DELETED status to a BatchPrediction, rendering it unusable.

After using the DeleteBatchPrediction operation, you can use the GetBatchPrediction operation to verify that the status of the BatchPrediction changed to DELETED.

Caution The result of the DeleteBatchPrediction operation is irreversible.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_DeleteBatchPrediction.html

Request

data DeleteBatchPrediction

Request constructor

deleteBatchPrediction

Arguments

:: Text
dbpBatchPredictionId

-> DeleteBatchPrediction

DeleteBatchPrediction constructor.

The fields accessible through corresponding lenses are:

dbpBatchPredictionId :: Text

Request lenses

dbpBatchPredictionId :: Lens' DeleteBatchPrediction Text

A user-supplied ID that uniquely identifies the BatchPrediction.

Response

data DeleteBatchPredictionResponse

Response constructor

deleteBatchPredictionResponse :: DeleteBatchPredictionResponse

DeleteBatchPredictionResponse constructor.

The fields accessible through corresponding lenses are:

dbprBatchPredictionId :: Maybe Text

Response lenses

dbprBatchPredictionId :: Lens' DeleteBatchPredictionResponse (Maybe Text)

A user-supplied ID that uniquely identifies the BatchPrediction. This value should be identical to the value of the BatchPredictionID in the request.