Network.AWS.MachineLearning.DeleteDataSource

Description

Assigns the DELETED status to a DataSource, rendering it unusable.

After using the DeleteDataSource operation, you can use the GetDataSource operation to verify that the status of the DataSource changed to DELETED.

Caution The results of the DeleteDataSource operation are irreversible.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_DeleteDataSource.html

Request

data DeleteDataSource

Request constructor

deleteDataSource

Arguments

:: Text
ddsDataSourceId

-> DeleteDataSource

DeleteDataSource constructor.

The fields accessible through corresponding lenses are:

ddsDataSourceId :: Text

Request lenses

ddsDataSourceId :: Lens' DeleteDataSource Text

A user-supplied ID that uniquely identifies the DataSource.

Response

data DeleteDataSourceResponse

Response constructor

deleteDataSourceResponse :: DeleteDataSourceResponse

DeleteDataSourceResponse constructor.

The fields accessible through corresponding lenses are:

ddsrDataSourceId :: Maybe Text

Response lenses

ddsrDataSourceId :: Lens' DeleteDataSourceResponse (Maybe Text)

A user-supplied ID that uniquely identifies the DataSource. This value should be identical to the value of the DataSourceID in the request.