Network.AWS.MachineLearning.DeleteEvaluation

Description

Assigns the DELETED status to an Evaluation, rendering it unusable.

After invoking the DeleteEvaluation operation, you can use the GetEvaluation operation to verify that the status of the Evaluation changed to DELETED.

Caution The results of the DeleteEvaluation operation are irreversible.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_DeleteEvaluation.html

Request

data DeleteEvaluation

Request constructor

deleteEvaluation

Arguments

:: Text
deEvaluationId

-> DeleteEvaluation

DeleteEvaluation constructor.

The fields accessible through corresponding lenses are:

deEvaluationId :: Text

Request lenses

deEvaluationId :: Lens' DeleteEvaluation Text

A user-supplied ID that uniquely identifies the Evaluation to delete.

Response

data DeleteEvaluationResponse

Response constructor

deleteEvaluationResponse :: DeleteEvaluationResponse

DeleteEvaluationResponse constructor.

The fields accessible through corresponding lenses are:

derEvaluationId :: Maybe Text

Response lenses

derEvaluationId :: Lens' DeleteEvaluationResponse (Maybe Text)

A user-supplied ID that uniquely identifies the Evaluation. This value should be identical to the value of the EvaluationId in the request.