Network.AWS.MachineLearning.DescribeBatchPredictions

Description

Returns a list of BatchPrediction operations that match the search criteria in the request.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_DescribeBatchPredictions.html

Request

data DescribeBatchPredictions

Request constructor

describeBatchPredictions :: DescribeBatchPredictions

DescribeBatchPredictions constructor.

The fields accessible through corresponding lenses are:

dbpEQ :: Maybe Text

dbpFilterVariable :: Maybe BatchPredictionFilterVariable

dbpGE :: Maybe Text

dbpGT :: Maybe Text

dbpLE :: Maybe Text

dbpLT :: Maybe Text

dbpLimit :: Maybe Natural

dbpNE :: Maybe Text

dbpNextToken :: Maybe Text

dbpPrefix :: Maybe Text

dbpSortOrder :: Maybe SortOrder

Request lenses

dbpEQ :: Lens' DescribeBatchPredictions (Maybe Text)

The equal to operator. The BatchPrediction results will have FilterVariable values that exactly match the value specified with EQ.

dbpFilterVariable :: Lens' DescribeBatchPredictions (Maybe BatchPredictionFilterVariable)

Use one of the following variables to filter a list of BatchPrediction:

CreatedAt - Sets the search criteria to the BatchPrediction creation date. Status - Sets the search criteria to the BatchPrediction status. Name - Sets the search criteria to the contents of the BatchPrediction Name. IAMUser - Sets the search criteria to the user account that invoked the BatchPrediction creation. MLModelId - Sets the search criteria to the MLModel used in the BatchPrediction. DataSourceId - Sets the search criteria to the DataSource used in the BatchPrediction. DataURI - Sets the search criteria to the data file(s) used in the BatchPrediction. The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

dbpGE :: Lens' DescribeBatchPredictions (Maybe Text)

The greater than or equal to operator. The BatchPrediction results will have FilterVariable values that are greater than or equal to the value specified with GE.

dbpGT :: Lens' DescribeBatchPredictions (Maybe Text)

The greater than operator. The BatchPrediction results will have FilterVariable values that are greater than the value specified with GT.

dbpLE :: Lens' DescribeBatchPredictions (Maybe Text)

The less than or equal to operator. The BatchPrediction results will have FilterVariable values that are less than or equal to the value specified with LE.

dbpLT :: Lens' DescribeBatchPredictions (Maybe Text)

The less than operator. The BatchPrediction results will have FilterVariable values that are less than the value specified with LT.

dbpLimit :: Lens' DescribeBatchPredictions (Maybe Natural)

The number of pages of information to include in the result. The range of acceptable values is 1 through 100. The default value is 100.

dbpNE :: Lens' DescribeBatchPredictions (Maybe Text)

The not equal to operator. The BatchPrediction results will have FilterVariable values not equal to the value specified with NE.

dbpNextToken :: Lens' DescribeBatchPredictions (Maybe Text)

An ID of the page in the paginated results.

dbpPrefix :: Lens' DescribeBatchPredictions (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id.

For example, a 'Batch Prediction' operation could have the Name '2014-09-09-HolidayGiftMailer'. To search for this BatchPrediction, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

dbpSortOrder :: Lens' DescribeBatchPredictions (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of MLModels.

asc - Arranges the list in ascending order (A-Z, 0-9). dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable.

Response

data DescribeBatchPredictionsResponse

Response constructor

describeBatchPredictionsResponse :: DescribeBatchPredictionsResponse

DescribeBatchPredictionsResponse constructor.

The fields accessible through corresponding lenses are:

dbprNextToken :: Maybe Text

dbprResults :: [BatchPrediction]

Response lenses

dbprNextToken :: Lens' DescribeBatchPredictionsResponse (Maybe Text)

The ID of the next page in the paginated results that indicates at least one more page follows.

dbprResults :: Lens' DescribeBatchPredictionsResponse [BatchPrediction]

A list of BatchPrediction objects that meet the search criteria.