Network.AWS.MachineLearning.DescribeDataSources

Description

Returns a list of DataSource that match the search criteria in the request.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_DescribeDataSources.html

Request

data DescribeDataSources

Request constructor

describeDataSources :: DescribeDataSources

DescribeDataSources constructor.

The fields accessible through corresponding lenses are:

ddsEQ :: Maybe Text

ddsFilterVariable :: Maybe DataSourceFilterVariable

ddsGE :: Maybe Text

ddsGT :: Maybe Text

ddsLE :: Maybe Text

ddsLT :: Maybe Text

ddsLimit :: Maybe Natural

ddsNE :: Maybe Text

ddsNextToken :: Maybe Text

ddsPrefix :: Maybe Text

ddsSortOrder :: Maybe SortOrder

Request lenses

ddsEQ :: Lens' DescribeDataSources (Maybe Text)

The equal to operator. The DataSource results will have FilterVariable values that exactly match the value specified with EQ.

ddsFilterVariable :: Lens' DescribeDataSources (Maybe DataSourceFilterVariable)

Use one of the following variables to filter a list of DataSource:

CreatedAt - Sets the search criteria to DataSource creation dates. Status - Sets the search criteria to DataSource statuses. Name - Sets the search criteria to the contents of DataSource Name. DataUri - Sets the search criteria to the URI of data files used to create the DataSource. The URI can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory. IAMUser - Sets the search criteria to the user account that invoked the DataSource creation.

ddsGE :: Lens' DescribeDataSources (Maybe Text)

The greater than or equal to operator. The DataSource results will have FilterVariable values that are greater than or equal to the value specified with GE.

ddsGT :: Lens' DescribeDataSources (Maybe Text)

The greater than operator. The DataSource results will have FilterVariable values that are greater than the value specified with GT.

ddsLE :: Lens' DescribeDataSources (Maybe Text)

The less than or equal to operator. The DataSource results will have FilterVariable values that are less than or equal to the value specified with LE.

ddsLT :: Lens' DescribeDataSources (Maybe Text)

The less than operator. The DataSource results will have FilterVariable values that are less than the value specified with LT.

ddsLimit :: Lens' DescribeDataSources (Maybe Natural)

The maximum number of DataSource to include in the result.

ddsNE :: Lens' DescribeDataSources (Maybe Text)

The not equal to operator. The DataSource results will have FilterVariable values not equal to the value specified with NE.

ddsNextToken :: Lens' DescribeDataSources (Maybe Text)

The ID of the page in the paginated results.

ddsPrefix :: Lens' DescribeDataSources (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id.

For example, a DataSource could have the Name '2014-09-09-HolidayGiftMailer'. To search for this DataSource, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

ddsSortOrder :: Lens' DescribeDataSources (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of DataSource.

asc - Arranges the list in ascending order (A-Z, 0-9). dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable.

Response

data DescribeDataSourcesResponse

Response constructor

describeDataSourcesResponse :: DescribeDataSourcesResponse

DescribeDataSourcesResponse constructor.

The fields accessible through corresponding lenses are:

ddsrNextToken :: Maybe Text

ddsrResults :: [DataSource]

Response lenses

ddsrNextToken :: Lens' DescribeDataSourcesResponse (Maybe Text)

An ID of the next page in the paginated results that indicates at least one more page follows.

ddsrResults :: Lens' DescribeDataSourcesResponse [DataSource]

A list of DataSource that meet the search criteria.