Network.AWS.MachineLearning.DescribeEvaluations

Description

Returns a list of DescribeEvaluations that match the search criteria in the request.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_DescribeEvaluations.html

Request

data DescribeEvaluations

Request constructor

describeEvaluations :: DescribeEvaluations

DescribeEvaluations constructor.

The fields accessible through corresponding lenses are:

deEQ :: Maybe Text

deFilterVariable :: Maybe EvaluationFilterVariable

deGE :: Maybe Text

deGT :: Maybe Text

deLE :: Maybe Text

deLT :: Maybe Text

deLimit :: Maybe Natural

deNE :: Maybe Text

deNextToken :: Maybe Text

dePrefix :: Maybe Text

deSortOrder :: Maybe SortOrder

Request lenses

deEQ :: Lens' DescribeEvaluations (Maybe Text)

The equal to operator. The Evaluation results will have FilterVariable values that exactly match the value specified with EQ.

deFilterVariable :: Lens' DescribeEvaluations (Maybe EvaluationFilterVariable)

Use one of the following variable to filter a list of Evaluation objects:

CreatedAt - Sets the search criteria to the Evaluation creation date. Status - Sets the search criteria to the Evaluation status. Name - Sets the search criteria to the contents of Evaluation Name. IAMUser - Sets the search criteria to the user account that invoked an Evaluation. MLModelId - Sets the search criteria to the MLModel that was evaluated. DataSourceId - Sets the search criteria to the DataSource used in Evaluation. DataUri - Sets the search criteria to the data file(s) used in Evaluation. The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

deGE :: Lens' DescribeEvaluations (Maybe Text)

The greater than or equal to operator. The Evaluation results will have FilterVariable values that are greater than or equal to the value specified with GE.

deGT :: Lens' DescribeEvaluations (Maybe Text)

The greater than operator. The Evaluation results will have FilterVariable values that are greater than the value specified with GT.

deLE :: Lens' DescribeEvaluations (Maybe Text)

The less than or equal to operator. The Evaluation results will have FilterVariable values that are less than or equal to the value specified with LE.

deLT :: Lens' DescribeEvaluations (Maybe Text)

The less than operator. The Evaluation results will have FilterVariable values that are less than the value specified with LT.

deLimit :: Lens' DescribeEvaluations (Maybe Natural)

The maximum number of Evaluation to include in the result.

deNE :: Lens' DescribeEvaluations (Maybe Text)

The not equal to operator. The Evaluation results will have FilterVariable values not equal to the value specified with NE.

deNextToken :: Lens' DescribeEvaluations (Maybe Text)

The ID of the page in the paginated results.

dePrefix :: Lens' DescribeEvaluations (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id.

For example, an Evaluation could have the Name '2014-09-09-HolidayGiftMailer'. To search for this Evaluation, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

deSortOrder :: Lens' DescribeEvaluations (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of Evaluation.

asc - Arranges the list in ascending order (A-Z, 0-9). dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable.

Response

data DescribeEvaluationsResponse

Response constructor

describeEvaluationsResponse :: DescribeEvaluationsResponse

DescribeEvaluationsResponse constructor.

The fields accessible through corresponding lenses are:

derNextToken :: Maybe Text

derResults :: [Evaluation]

Response lenses

derNextToken :: Lens' DescribeEvaluationsResponse (Maybe Text)

The ID of the next page in the paginated results that indicates at least one more page follows.

derResults :: Lens' DescribeEvaluationsResponse [Evaluation]

A list of Evaluation that meet the search criteria.