Network.AWS.MachineLearning.DescribeMLModels

Description

Returns a list of MLModel that match the search criteria in the request.

This operation returns paginated results.

Creating a Request

describeMLModels :: DescribeMLModels

Creates a value of DescribeMLModels with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmlmEQ - The equal to operator. The MLModel results will have FilterVariable values that exactly match the value specified with EQ .

dmlmGE - The greater than or equal to operator. The MLModel results will have FilterVariable values that are greater than or equal to the value specified with GE .

dmlmPrefix - A string that is found at the beginning of a variable, such as Name or Id . For example, an MLModel could have the Name 2014-09-09-HolidayGiftMailer . To search for this MLModel , select Name for the FilterVariable and any of the following strings for the Prefix : * 2014-09 * 2014-09-09 * 2014-09-09-Holiday

dmlmGT - The greater than operator. The MLModel results will have FilterVariable values that are greater than the value specified with GT .

dmlmNE - The not equal to operator. The MLModel results will have FilterVariable values not equal to the value specified with NE .

dmlmNextToken - The ID of the page in the paginated results.

dmlmSortOrder - A two-value parameter that determines the sequence of the resulting list of MLModel . * asc - Arranges the list in ascending order (A-Z, 0-9). * dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable .

dmlmLimit - The number of pages of information to include in the result. The range of acceptable values is 1 through 100 . The default value is 100 .

dmlmLT - The less than operator. The MLModel results will have FilterVariable values that are less than the value specified with LT .

dmlmFilterVariable - Use one of the following variables to filter a list of MLModel : * CreatedAt - Sets the search criteria to MLModel creation date. * Status - Sets the search criteria to MLModel status. * Name - Sets the search criteria to the contents of MLModel ____ Name . * IAMUser - Sets the search criteria to the user account that invoked the MLModel creation. * TrainingDataSourceId - Sets the search criteria to the DataSource used to train one or more MLModel . * RealtimeEndpointStatus - Sets the search criteria to the MLModel real-time endpoint status. * MLModelType - Sets the search criteria to MLModel type: binary, regression, or multi-class. * Algorithm - Sets the search criteria to the algorithm that the MLModel uses. * TrainingDataURI - Sets the search criteria to the data file(s) used in training a MLModel . The URL can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

dmlmLE - The less than or equal to operator. The MLModel results will have FilterVariable values that are less than or equal to the value specified with LE .

data DescribeMLModels

See: describeMLModels smart constructor.

Request Lenses

dmlmEQ :: Lens' DescribeMLModels (Maybe Text)

The equal to operator. The MLModel results will have FilterVariable values that exactly match the value specified with EQ .

dmlmGE :: Lens' DescribeMLModels (Maybe Text)

The greater than or equal to operator. The MLModel results will have FilterVariable values that are greater than or equal to the value specified with GE .

dmlmPrefix :: Lens' DescribeMLModels (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id . For example, an MLModel could have the Name 2014-09-09-HolidayGiftMailer . To search for this MLModel , select Name for the FilterVariable and any of the following strings for the Prefix : * 2014-09 * 2014-09-09 * 2014-09-09-Holiday

dmlmGT :: Lens' DescribeMLModels (Maybe Text)

The greater than operator. The MLModel results will have FilterVariable values that are greater than the value specified with GT .

dmlmNE :: Lens' DescribeMLModels (Maybe Text)

The not equal to operator. The MLModel results will have FilterVariable values not equal to the value specified with NE .

dmlmNextToken :: Lens' DescribeMLModels (Maybe Text)

The ID of the page in the paginated results.

dmlmSortOrder :: Lens' DescribeMLModels (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of MLModel . * asc - Arranges the list in ascending order (A-Z, 0-9). * dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable .

dmlmLimit :: Lens' DescribeMLModels (Maybe Natural)

The number of pages of information to include in the result. The range of acceptable values is 1 through 100 . The default value is 100 .

dmlmLT :: Lens' DescribeMLModels (Maybe Text)

The less than operator. The MLModel results will have FilterVariable values that are less than the value specified with LT .

dmlmFilterVariable :: Lens' DescribeMLModels (Maybe MLModelFilterVariable)

Use one of the following variables to filter a list of MLModel : * CreatedAt - Sets the search criteria to MLModel creation date. * Status - Sets the search criteria to MLModel status. * Name - Sets the search criteria to the contents of MLModel ____ Name . * IAMUser - Sets the search criteria to the user account that invoked the MLModel creation. * TrainingDataSourceId - Sets the search criteria to the DataSource used to train one or more MLModel . * RealtimeEndpointStatus - Sets the search criteria to the MLModel real-time endpoint status. * MLModelType - Sets the search criteria to MLModel type: binary, regression, or multi-class. * Algorithm - Sets the search criteria to the algorithm that the MLModel uses. * TrainingDataURI - Sets the search criteria to the data file(s) used in training a MLModel . The URL can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

dmlmLE :: Lens' DescribeMLModels (Maybe Text)

The less than or equal to operator. The MLModel results will have FilterVariable values that are less than or equal to the value specified with LE .

Destructuring the Response

describeMLModelsResponse

Arguments

:: Int
dmlmsrsResponseStatus

-> DescribeMLModelsResponse

Creates a value of DescribeMLModelsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmlmsrsResults - A list of MLModel that meet the search criteria.

dmlmsrsNextToken - The ID of the next page in the paginated results that indicates at least one more page follows.

dmlmsrsResponseStatus - -- | The response status code.

data DescribeMLModelsResponse

Represents the output of a DescribeMLModels operation. The content is essentially a list of MLModel .

See: describeMLModelsResponse smart constructor.

Response Lenses

dmlmsrsResults :: Lens' DescribeMLModelsResponse [MLModel]

A list of MLModel that meet the search criteria.

dmlmsrsNextToken :: Lens' DescribeMLModelsResponse (Maybe Text)

The ID of the next page in the paginated results that indicates at least one more page follows.

dmlmsrsResponseStatus :: Lens' DescribeMLModelsResponse Int

- | The response status code.