Network.AWS.MachineLearning.GetBatchPrediction

Description

Returns a BatchPrediction that includes detailed metadata, status, and data file information for a 'Batch Prediction' request.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_GetBatchPrediction.html

Request

data GetBatchPrediction

Request constructor

getBatchPrediction

Arguments

:: Text
gbpBatchPredictionId

-> GetBatchPrediction

GetBatchPrediction constructor.

The fields accessible through corresponding lenses are:

gbpBatchPredictionId :: Text

Request lenses

gbpBatchPredictionId :: Lens' GetBatchPrediction Text

An ID assigned to the BatchPrediction at creation.

Response

data GetBatchPredictionResponse

Response constructor

getBatchPredictionResponse :: GetBatchPredictionResponse

GetBatchPredictionResponse constructor.

The fields accessible through corresponding lenses are:

gbprBatchPredictionDataSourceId :: Maybe Text

gbprBatchPredictionId :: Maybe Text

gbprCreatedAt :: Maybe UTCTime

gbprCreatedByIamUser :: Maybe Text

gbprInputDataLocationS3 :: Maybe Text

gbprLastUpdatedAt :: Maybe UTCTime

gbprLogUri :: Maybe Text

gbprMLModelId :: Maybe Text

gbprMessage :: Maybe Text

gbprName :: Maybe Text

gbprOutputUri :: Maybe Text

gbprStatus :: Maybe EntityStatus

Response lenses

gbprBatchPredictionDataSourceId :: Lens' GetBatchPredictionResponse (Maybe Text)

The ID of the DataSource that was used to create the BatchPrediction.

gbprBatchPredictionId :: Lens' GetBatchPredictionResponse (Maybe Text)

An ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

gbprCreatedAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The time when the BatchPrediction was created. The time is expressed in epoch time.

gbprCreatedByIamUser :: Lens' GetBatchPredictionResponse (Maybe Text)

The AWS user account that invoked the BatchPrediction. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gbprInputDataLocationS3 :: Lens' GetBatchPredictionResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gbprLastUpdatedAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The time of the most recent edit to BatchPrediction. The time is expressed in epoch time.

gbprLogUri :: Lens' GetBatchPredictionResponse (Maybe Text)

A link to the file that contains logs of the CreateBatchPrediction operation.

gbprMLModelId :: Lens' GetBatchPredictionResponse (Maybe Text)

The ID of the MLModel that generated predictions for the BatchPrediction request.

gbprMessage :: Lens' GetBatchPredictionResponse (Maybe Text)

A description of the most recent details about processing the batch prediction request.

gbprName :: Lens' GetBatchPredictionResponse (Maybe Text)

A user-supplied name or description of the BatchPrediction.

gbprOutputUri :: Lens' GetBatchPredictionResponse (Maybe Text)

The location of an Amazon S3 bucket or directory to receive the operation results.

gbprStatus :: Lens' GetBatchPredictionResponse (Maybe EntityStatus)

The status of the BatchPrediction, which can be one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate batch predictions. INPROGRESS - The batch predictions are in progress. FAILED - The request to perform a batch prediction did not run to completion. It is not usable. COMPLETED - The batch prediction process completed successfully. DELETED - The BatchPrediction is marked as deleted. It is not usable.