Network.AWS.MachineLearning.GetDataSource

Description

Returns a DataSource that includes metadata and data file information, as well as the current status of the DataSource.

GetDataSource provides results in normal or verbose format. The verbose format adds the schema description and the list of files pointed to by the DataSource to the normal format.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_GetDataSource.html

Request

data GetDataSource

Request constructor

getDataSource

Arguments

:: Text
gdsDataSourceId

-> GetDataSource

GetDataSource constructor.

The fields accessible through corresponding lenses are:

gdsDataSourceId :: Text

gdsVerbose :: Maybe Bool

Request lenses

gdsDataSourceId :: Lens' GetDataSource Text

The ID assigned to the DataSource at creation.

gdsVerbose :: Lens' GetDataSource (Maybe Bool)

Specifies whether the GetDataSource operation should return DataSourceSchema.

If true, DataSourceSchema is returned.

If false, DataSourceSchema is not returned.

Response

data GetDataSourceResponse

Response constructor

getDataSourceResponse :: GetDataSourceResponse

GetDataSourceResponse constructor.

The fields accessible through corresponding lenses are:

gdsrComputeStatistics :: Maybe Bool

gdsrCreatedAt :: Maybe UTCTime

gdsrCreatedByIamUser :: Maybe Text

gdsrDataLocationS3 :: Maybe Text

gdsrDataRearrangement :: Maybe Text

gdsrDataSizeInBytes :: Maybe Integer

gdsrDataSourceId :: Maybe Text

gdsrDataSourceSchema :: Maybe Text

gdsrLastUpdatedAt :: Maybe UTCTime

gdsrLogUri :: Maybe Text

gdsrMessage :: Maybe Text

gdsrName :: Maybe Text

gdsrNumberOfFiles :: Maybe Integer

gdsrRDSMetadata :: Maybe RDSMetadata

gdsrRedshiftMetadata :: Maybe RedshiftMetadata

gdsrRoleARN :: Maybe Text

gdsrStatus :: Maybe EntityStatus

Response lenses

gdsrComputeStatistics :: Lens' GetDataSourceResponse (Maybe Bool)

The parameter is true if statistics need to be generated from the observation data.

gdsrCreatedAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The time that the DataSource was created. The time is expressed in epoch time.

gdsrCreatedByIamUser :: Lens' GetDataSourceResponse (Maybe Text)

The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gdsrDataLocationS3 :: Lens' GetDataSourceResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gdsrDataRearrangement :: Lens' GetDataSourceResponse (Maybe Text)

A JSON string that captures the splitting rearrangement requirement of the DataSource.

gdsrDataSizeInBytes :: Lens' GetDataSourceResponse (Maybe Integer)

The total size of observations in the data files.

gdsrDataSourceId :: Lens' GetDataSourceResponse (Maybe Text)

The ID assigned to the DataSource at creation. This value should be identical to the value of the DataSourceId in the request.

gdsrDataSourceSchema :: Lens' GetDataSourceResponse (Maybe Text)

The schema used by all of the data files of this DataSource.

Note This parameter is provided as part of the verbose format.

gdsrLastUpdatedAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The time of the most recent edit to the DataSource. The time is expressed in epoch time.

gdsrLogUri :: Lens' GetDataSourceResponse (Maybe Text)

A link to the file containining logs of either create DataSource operation.

gdsrMessage :: Lens' GetDataSourceResponse (Maybe Text)

The description of the most recent details about creating the DataSource.

gdsrName :: Lens' GetDataSourceResponse (Maybe Text)

A user-supplied name or description of the DataSource.

gdsrNumberOfFiles :: Lens' GetDataSourceResponse (Maybe Integer)

The number of data files referenced by the DataSource.

gdsrRDSMetadata :: Lens' GetDataSourceResponse (Maybe RDSMetadata)

gdsrRedshiftMetadata :: Lens' GetDataSourceResponse (Maybe RedshiftMetadata)

gdsrRoleARN :: Lens' GetDataSourceResponse (Maybe Text)

gdsrStatus :: Lens' GetDataSourceResponse (Maybe EntityStatus)

The current status of the DataSource. This element can have one of the following values:

PENDING - Amazon Machine Language (Amazon ML) submitted a request to create a DataSource. INPROGRESS - The creation process is underway. FAILED - The request to create a DataSource did not run to completion. It is not usable. COMPLETED - The creation process completed successfully. DELETED - The DataSource is marked as deleted. It is not usable.