Network.AWS.MachineLearning.GetEvaluation

Description

Returns an Evaluation that includes metadata as well as the current status of the Evaluation.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_GetEvaluation.html

Request

data GetEvaluation

Request constructor

getEvaluation

Arguments

:: Text
geEvaluationId

-> GetEvaluation

GetEvaluation constructor.

The fields accessible through corresponding lenses are:

geEvaluationId :: Text

Request lenses

geEvaluationId :: Lens' GetEvaluation Text

The ID of the Evaluation to retrieve. The evaluation of each MLModel is recorded and cataloged. The ID provides the means to access the information.

Response

data GetEvaluationResponse

Response constructor

getEvaluationResponse :: GetEvaluationResponse

GetEvaluationResponse constructor.

The fields accessible through corresponding lenses are:

gerCreatedAt :: Maybe UTCTime

gerCreatedByIamUser :: Maybe Text

gerEvaluationDataSourceId :: Maybe Text

gerEvaluationId :: Maybe Text

gerInputDataLocationS3 :: Maybe Text

gerLastUpdatedAt :: Maybe UTCTime

gerLogUri :: Maybe Text

gerMLModelId :: Maybe Text

gerMessage :: Maybe Text

gerName :: Maybe Text

gerPerformanceMetrics :: Maybe PerformanceMetrics

gerStatus :: Maybe EntityStatus

Response lenses

gerCreatedAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The time that the Evaluation was created. The time is expressed in epoch time.

gerCreatedByIamUser :: Lens' GetEvaluationResponse (Maybe Text)

The AWS user account that invoked the evaluation. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gerEvaluationDataSourceId :: Lens' GetEvaluationResponse (Maybe Text)

The DataSource used for this evaluation.

gerEvaluationId :: Lens' GetEvaluationResponse (Maybe Text)

The evaluation ID which is same as the EvaluationId in the request.

gerInputDataLocationS3 :: Lens' GetEvaluationResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gerLastUpdatedAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.

gerLogUri :: Lens' GetEvaluationResponse (Maybe Text)

A link to the file that contains logs of the CreateEvaluation operation.

gerMLModelId :: Lens' GetEvaluationResponse (Maybe Text)

The ID of the MLModel that was the focus of the evaluation.

gerMessage :: Lens' GetEvaluationResponse (Maybe Text)

A description of the most recent details about evaluating the MLModel.

gerName :: Lens' GetEvaluationResponse (Maybe Text)

A user-supplied name or description of the Evaluation.

gerPerformanceMetrics :: Lens' GetEvaluationResponse (Maybe PerformanceMetrics)

Measurements of how well the MLModel performed using observations referenced by the DataSource. One of the following metric is returned based on the type of the MLModel:

BinaryAUC: A binary MLModel uses the Area Under the Curve (AUC) technique to measure performance.

RegressionRMSE: A regression MLModel uses the Root Mean Square Error (RMSE) technique to measure performance. RMSE measures the difference between predicted and actual values for a single variable.

MulticlassAvgFScore: A multiclass MLModel uses the F1 score technique to measure performance.

For more information about performance metrics, please see the AmazonMachine Learning Developer Guide.

gerStatus :: Lens' GetEvaluationResponse (Maybe EntityStatus)

The status of the evaluation. This element can have one of the following values:

PENDING - Amazon Machine Language (Amazon ML) submitted a request to evaluate an MLModel. INPROGRESS - The evaluation is underway. FAILED - The request to evaluate an MLModel did not run to completion. It is not usable. COMPLETED - The evaluation process completed successfully. DELETED - The Evaluation is marked as deleted. It is not usable.