Network.AWS.MachineLearning.GetMLModel

Description

Returns an MLModel that includes detailed metadata, and data source information as well as the current status of the MLModel.

GetMLModel provides results in normal or verbose format.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_GetMLModel.html

Request

data GetMLModel

Request constructor

getMLModel

Arguments

:: Text
gmlmMLModelId

-> GetMLModel

GetMLModel constructor.

The fields accessible through corresponding lenses are:

gmlmMLModelId :: Text

gmlmVerbose :: Maybe Bool

Request lenses

gmlmMLModelId :: Lens' GetMLModel Text

The ID assigned to the MLModel at creation.

gmlmVerbose :: Lens' GetMLModel (Maybe Bool)

Specifies whether the GetMLModel operation should return Recipe.

If true, Recipe is returned.

If false, Recipe is not returned.

Response

data GetMLModelResponse

Response constructor

getMLModelResponse :: GetMLModelResponse

GetMLModelResponse constructor.

The fields accessible through corresponding lenses are:

gmlmrCreatedAt :: Maybe UTCTime

gmlmrCreatedByIamUser :: Maybe Text

gmlmrEndpointInfo :: Maybe RealtimeEndpointInfo

gmlmrInputDataLocationS3 :: Maybe Text

gmlmrLastUpdatedAt :: Maybe UTCTime

gmlmrLogUri :: Maybe Text

gmlmrMLModelId :: Maybe Text

gmlmrMLModelType :: Maybe MLModelType

gmlmrMessage :: Maybe Text

gmlmrName :: Maybe Text

gmlmrRecipe :: Maybe Text

gmlmrSchema :: Maybe Text

gmlmrScoreThreshold :: Maybe Double

gmlmrScoreThresholdLastUpdatedAt :: Maybe UTCTime

gmlmrSizeInBytes :: Maybe Integer

gmlmrStatus :: Maybe EntityStatus

gmlmrTrainingDataSourceId :: Maybe Text

gmlmrTrainingParameters :: HashMap Text Text

Response lenses

gmlmrCreatedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The time that the MLModel was created. The time is expressed in epoch time.

gmlmrCreatedByIamUser :: Lens' GetMLModelResponse (Maybe Text)

The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gmlmrEndpointInfo :: Lens' GetMLModelResponse (Maybe RealtimeEndpointInfo)

The current endpoint of the MLModel

gmlmrInputDataLocationS3 :: Lens' GetMLModelResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gmlmrLastUpdatedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The time of the most recent edit to the MLModel. The time is expressed in epoch time.

gmlmrLogUri :: Lens' GetMLModelResponse (Maybe Text)

A link to the file that contains logs of the CreateMLModel operation.

gmlmrMLModelId :: Lens' GetMLModelResponse (Maybe Text)

The MLModel ID which is same as the MLModelId in the request.

gmlmrMLModelType :: Lens' GetMLModelResponse (Maybe MLModelType)

Identifies the MLModel category. The following are the available types:

REGRESSION -- Produces a numeric result. For example, "What listing price should a house have?" BINARY -- Produces one of two possible results. For example, "Is this an e-commerce website?" MULTICLASS -- Produces more than two possible results. For example, "Is this a HIGH, LOW or MEDIUM risk trade?"

gmlmrMessage :: Lens' GetMLModelResponse (Maybe Text)

Description of the most recent details about accessing the MLModel.

gmlmrName :: Lens' GetMLModelResponse (Maybe Text)

A user-supplied name or description of the MLModel.

gmlmrRecipe :: Lens' GetMLModelResponse (Maybe Text)

The recipe to use when training the MLModel. The Recipe provides detailed information about the observation data to use during training, as well as manipulations to perform on the observation data during training.

Note This parameter is provided as part of the verbose format.

gmlmrSchema :: Lens' GetMLModelResponse (Maybe Text)

The schema used by all of the data files referenced by the DataSource.

Note This parameter is provided as part of the verbose format.

gmlmrScoreThreshold :: Lens' GetMLModelResponse (Maybe Double)

The scoring threshold is used in binary classification MLModels, and marks the boundary between a positive prediction and a negative prediction.

Output values greater than or equal to the threshold receive a positive result from the MLModel, such as true. Output values less than the threshold receive a negative response from the MLModel, such as false.

gmlmrScoreThresholdLastUpdatedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

gmlmrSizeInBytes :: Lens' GetMLModelResponse (Maybe Integer)

gmlmrStatus :: Lens' GetMLModelResponse (Maybe EntityStatus)

The current status of the MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to describe a MLModel. INPROGRESS - The request is processing. FAILED - The request did not run to completion. It is not usable. COMPLETED - The request completed successfully. DELETED - The MLModel is marked as deleted. It is not usable.

gmlmrTrainingDataSourceId :: Lens' GetMLModelResponse (Maybe Text)

The ID of the training DataSource.

gmlmrTrainingParameters :: Lens' GetMLModelResponse (HashMap Text Text)

A list of the training parameters in the MLModel. The list is implemented as a map of key/value pairs.

The following is the current set of training parameters:

'sgd.l1RegularizationAmount' - Coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, specify a small value, such as 1.0E-04 or 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is not to use L1 normalization. The parameter cannot be used when L2 is specified. Use this parameter sparingly.

'sgd.l2RegularizationAmount' - Coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, specify a small value, such as 1.0E-04 or 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is not to use L2 normalization. This parameter cannot be used when L1 is specified. Use this parameter sparingly.

'sgd.maxPasses' - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

'sgd.maxMLModelSizeInBytes' - The maximum allowed size of the model. Depending on the input data, the model size might affect performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.