Network.AWS.MachineLearning.Predict

Description

Generates a prediction for the observation using the specified MLModel.

Note Not all response parameters will be populated because this is dependent on the type of requested model.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_Predict.html

Request

data Predict

Request constructor

predict

Arguments

:: Text
pMLModelId

-> Text
pPredictEndpoint

-> Predict

Predict constructor.

The fields accessible through corresponding lenses are:

pMLModelId :: Text

pPredictEndpoint :: Text

pRecord :: HashMap Text Text

Request lenses

pMLModelId :: Lens' Predict Text

A unique identifier of the MLModel.

pPredictEndpoint :: Lens' Predict Text

pRecord :: Lens' Predict (HashMap Text Text)

Response

data PredictResponse

Response constructor

predictResponse :: PredictResponse

PredictResponse constructor.

The fields accessible through corresponding lenses are:

prPrediction :: Maybe Prediction

Response lenses

prPrediction :: Lens' PredictResponse (Maybe Prediction)