Network.AWS.MachineLearning.Types

Service

data MachineLearning

Version 2014-12-12 of the Amazon Machine Learning service.

Error

data JSONError :: *

BatchPredictionFilterVariable

data BatchPredictionFilterVariable

Constructors

BatchCreatedAt
CreatedAt

BatchDataSourceId
DataSourceId

BatchDataURI
DataURI

BatchIAMUser
IAMUser

BatchLastUpdatedAt
LastUpdatedAt

BatchMLModelId
MLModelId

BatchName
Name

BatchStatus
Status

PerformanceMetrics

data PerformanceMetrics

performanceMetrics :: PerformanceMetrics

PerformanceMetrics constructor.

The fields accessible through corresponding lenses are:

pmProperties :: HashMap Text Text

pmProperties :: Lens' PerformanceMetrics (HashMap Text Text)

RealtimeEndpointInfo

data RealtimeEndpointInfo

realtimeEndpointInfo :: RealtimeEndpointInfo

RealtimeEndpointInfo constructor.

The fields accessible through corresponding lenses are:

reiCreatedAt :: Maybe UTCTime

reiEndpointStatus :: Maybe RealtimeEndpointStatus

reiEndpointUrl :: Maybe Text

reiPeakRequestsPerSecond :: Maybe Int

reiCreatedAt :: Lens' RealtimeEndpointInfo (Maybe UTCTime)

The time that the request to create the real-time endpoint for the MLModel was received. The time is expressed in epoch time.

reiEndpointStatus :: Lens' RealtimeEndpointInfo (Maybe RealtimeEndpointStatus)

The current status of the real-time endpoint for the MLModel. This element can have one of the following values:

NONE - Endpoint does not exist or was previously deleted. READY - Endpoint is ready to be used for real-time predictions. UPDATING - Updating/creating the endpoint.

reiEndpointUrl :: Lens' RealtimeEndpointInfo (Maybe Text)

The URI that specifies where to send real-time prediction requests for the MLModel.

Note The application must wait until the real-time endpoint is ready before using this URI.

reiPeakRequestsPerSecond :: Lens' RealtimeEndpointInfo (Maybe Int)

The maximum processing rate for the real-time endpoint for MLModel, measured in incoming requests per second.

Prediction

data Prediction

prediction :: Prediction

Prediction constructor.

The fields accessible through corresponding lenses are:

pDetails :: HashMap DetailsAttributes Text

pPredictedLabel :: Maybe Text

pPredictedScores :: HashMap Text Double

pPredictedValue :: Maybe Double

pDetails :: Lens' Prediction (HashMap DetailsAttributes Text)

pPredictedLabel :: Lens' Prediction (Maybe Text)

The prediction label for either a BINARY or MULTICLASS MLModel.

pPredictedScores :: Lens' Prediction (HashMap Text Double)

pPredictedValue :: Lens' Prediction (Maybe Double)

The prediction value for REGRESSION MLModel.

S3DataSpec

data S3DataSpec

s3DataSpec

Arguments

:: Text
sdsDataLocationS3

-> S3DataSpec

S3DataSpec constructor.

The fields accessible through corresponding lenses are:

sdsDataLocationS3 :: Text

sdsDataRearrangement :: Maybe Text

sdsDataSchema :: Maybe Text

sdsDataSchemaLocationS3 :: Maybe Text

sdsDataLocationS3 :: Lens' S3DataSpec Text

The location of the data file(s) used by a DataSource. The URI specifies a data file or an Amazon Simple Storage Service (Amazon S3) directory or bucket containing data files.

sdsDataRearrangement :: Lens' S3DataSpec (Maybe Text)

Describes the splitting requirement of a Datasource.

sdsDataSchema :: Lens' S3DataSpec (Maybe Text)

Describes the schema for an Amazon S3 DataSource.

sdsDataSchemaLocationS3 :: Lens' S3DataSpec (Maybe Text)

Describes the schema Location in Amazon S3.

DetailsAttributes

data DetailsAttributes

Constructors

DAAlgorithm
Algorithm

DAPredictiveModelType
PredictiveModelType

EvaluationFilterVariable

data EvaluationFilterVariable

Constructors

EvalCreatedAt
CreatedAt

EvalDataSourceId
DataSourceId

EvalDataURI
DataURI

EvalIAMUser
IAMUser

EvalLastUpdatedAt
LastUpdatedAt

EvalMLModelId
MLModelId

EvalName
Name

EvalStatus
Status

RealtimeEndpointStatus

data RealtimeEndpointStatus

Constructors

Failed
FAILED

None
NONE

Ready
READY

Updating
UPDATING

RDSDataSpec

data RDSDataSpec

rdsdataSpec

Arguments

:: RDSDatabase
rdsdsDatabaseInformation

-> Text
rdsdsSelectSqlQuery

-> RDSDatabaseCredentials
rdsdsDatabaseCredentials

-> Text
rdsdsS3StagingLocation

-> Text
rdsdsResourceRole

-> Text
rdsdsServiceRole

-> Text
rdsdsSubnetId

-> RDSDataSpec

RDSDataSpec constructor.

The fields accessible through corresponding lenses are:

rdsdsDataRearrangement :: Maybe Text

rdsdsDataSchema :: Maybe Text

rdsdsDataSchemaUri :: Maybe Text

rdsdsDatabaseCredentials :: RDSDatabaseCredentials

rdsdsDatabaseInformation :: RDSDatabase

rdsdsResourceRole :: Text

rdsdsS3StagingLocation :: Text

rdsdsSecurityGroupIds :: [Text]

rdsdsSelectSqlQuery :: Text

rdsdsServiceRole :: Text

rdsdsSubnetId :: Text

rdsdsDataRearrangement :: Lens' RDSDataSpec (Maybe Text)

DataRearrangement - A JSON string that represents the splitting requirement of a DataSource.

Sample - ' "{"randomSeed":"some-random-seed","splitting":{"percentBegin":10,"percentEnd":60}}"'

rdsdsDataSchema :: Lens' RDSDataSpec (Maybe Text)

A JSON string that represents the schema. This is not required if DataSchemaUri is specified.

rdsdsDataSchemaUri :: Lens' RDSDataSpec (Maybe Text)

The Amazon S3 location of the DataSchema.

rdsdsDatabaseCredentials :: Lens' RDSDataSpec RDSDatabaseCredentials

The AWS Identity and Access Management (IAM) credentials that are used connect to the Amazon RDS database.

rdsdsDatabaseInformation :: Lens' RDSDataSpec RDSDatabase

Describes the DatabaseName and InstanceIdentifier of an an Amazon RDS database.

rdsdsResourceRole :: Lens' RDSDataSpec Text

The role (DataPipelineDefaultResourceRole) assumed by an Amazon Elastic Compute Cloud (Amazon EC2) instance to carry out the copy operation from Amazon RDS to an Amazon S3 task. For more information, see Role templates for data pipelines.

rdsdsS3StagingLocation :: Lens' RDSDataSpec Text

The Amazon S3 location for staging Amazon RDS data. The data retrieved from Amazon RDS using SelectSqlQuery is stored in this location.

rdsdsSecurityGroupIds :: Lens' RDSDataSpec [Text]

The security group IDs to be used to access a VPC-based RDS DB instance. Ensure that there are appropriate ingress rules set up to allow access to the RDS DB instance. This attribute is used by Data Pipeline to carry out the copy operation from Amazon RDS to an Amazon S3 task.

rdsdsSelectSqlQuery :: Lens' RDSDataSpec Text

The query that is used to retrieve the observation data for the DataSource.

rdsdsServiceRole :: Lens' RDSDataSpec Text

The role (DataPipelineDefaultRole) assumed by AWS Data Pipeline service to monitor the progress of the copy task from Amazon RDS to Amazon S3. For more information, see Role templates for data pipelines.

rdsdsSubnetId :: Lens' RDSDataSpec Text

The subnet ID to be used to access a VPC-based RDS DB instance. This attribute is used by Data Pipeline to carry out the copy task from Amazon RDS to Amazon S3.

RDSMetadata

data RDSMetadata

rdsmetadata :: RDSMetadata

RDSMetadata constructor.

The fields accessible through corresponding lenses are:

rdsmDataPipelineId :: Maybe Text

rdsmDatabase :: Maybe RDSDatabase

rdsmDatabaseUserName :: Maybe Text

rdsmResourceRole :: Maybe Text

rdsmSelectSqlQuery :: Maybe Text

rdsmServiceRole :: Maybe Text

rdsmDataPipelineId :: Lens' RDSMetadata (Maybe Text)

The ID of the Data Pipeline instance that is used to carry to copy data from Amazon RDS to Amazon S3. You can use the ID to find details about the instance in the Data Pipeline console.

rdsmDatabase :: Lens' RDSMetadata (Maybe RDSDatabase)

The database details required to connect to an Amazon RDS.

rdsmDatabaseUserName :: Lens' RDSMetadata (Maybe Text)

rdsmResourceRole :: Lens' RDSMetadata (Maybe Text)

The role (DataPipelineDefaultResourceRole) assumed by an Amazon EC2 instance to carry out the copy task from Amazon RDS to Amazon S3. For more information, see Role templates for data pipelines.

rdsmSelectSqlQuery :: Lens' RDSMetadata (Maybe Text)

The SQL query that is supplied during CreateDataSourceFromRDS. Returns only if Verbose is true in GetDataSourceInput.

rdsmServiceRole :: Lens' RDSMetadata (Maybe Text)

The role (DataPipelineDefaultRole) assumed by the Data Pipeline service to monitor the progress of the copy task from Amazon RDS to Amazon S3. For more information, see Role templates for data pipelines.

RedshiftDatabase

data RedshiftDatabase

redshiftDatabase

Arguments

:: Text
rdDatabaseName

-> Text
rdClusterIdentifier

-> RedshiftDatabase

RedshiftDatabase constructor.

The fields accessible through corresponding lenses are:

rdClusterIdentifier :: Text

rdDatabaseName :: Text

rdClusterIdentifier :: Lens' RedshiftDatabase Text

rdDatabaseName :: Lens' RedshiftDatabase Text

RedshiftDatabaseCredentials

data RedshiftDatabaseCredentials

redshiftDatabaseCredentials

Arguments

:: Text
rdcUsername

-> Text
rdcPassword

-> RedshiftDatabaseCredentials

RedshiftDatabaseCredentials constructor.

The fields accessible through corresponding lenses are:

rdcPassword :: Text

rdcUsername :: Text

rdcPassword :: Lens' RedshiftDatabaseCredentials Text

rdcUsername :: Lens' RedshiftDatabaseCredentials Text

MLModel

data MLModel

mlmodel :: MLModel

MLModel constructor.

The fields accessible through corresponding lenses are:

mlmAlgorithm :: Maybe Algorithm

mlmCreatedAt :: Maybe UTCTime

mlmCreatedByIamUser :: Maybe Text

mlmEndpointInfo :: Maybe RealtimeEndpointInfo

mlmInputDataLocationS3 :: Maybe Text

mlmLastUpdatedAt :: Maybe UTCTime

mlmMLModelId :: Maybe Text

mlmMLModelType :: Maybe MLModelType

mlmMessage :: Maybe Text

mlmName :: Maybe Text

mlmScoreThreshold :: Maybe Double

mlmScoreThresholdLastUpdatedAt :: Maybe UTCTime

mlmSizeInBytes :: Maybe Integer

mlmStatus :: Maybe EntityStatus

mlmTrainingDataSourceId :: Maybe Text

mlmTrainingParameters :: HashMap Text Text

mlmAlgorithm :: Lens' MLModel (Maybe Algorithm)

The algorithm used to train the MLModel. The following algorithm is supported:

SGD -- Stochastic gradient descent. The goal of SGD is to minimize the gradient of the loss function.

mlmCreatedAt :: Lens' MLModel (Maybe UTCTime)

The time that the MLModel was created. The time is expressed in epoch time.

mlmCreatedByIamUser :: Lens' MLModel (Maybe Text)

The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

mlmEndpointInfo :: Lens' MLModel (Maybe RealtimeEndpointInfo)

The current endpoint of the MLModel.

mlmInputDataLocationS3 :: Lens' MLModel (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

mlmLastUpdatedAt :: Lens' MLModel (Maybe UTCTime)

The time of the most recent edit to the MLModel. The time is expressed in epoch time.

mlmMLModelId :: Lens' MLModel (Maybe Text)

The ID assigned to the MLModel at creation.

mlmMLModelType :: Lens' MLModel (Maybe MLModelType)

Identifies the MLModel category. The following are the available types:

REGRESSION - Produces a numeric result. For example, "What listing price should a house have?". BINARY - Produces one of two possible results. For example, "Is this a child-friendly web site?". MULTICLASS - Produces more than two possible results. For example, "Is this a HIGH, LOW or MEDIUM risk trade?".

mlmMessage :: Lens' MLModel (Maybe Text)

A description of the most recent details about accessing the MLModel.

mlmName :: Lens' MLModel (Maybe Text)

A user-supplied name or description of the MLModel.

mlmScoreThreshold :: Lens' MLModel (Maybe Double)

mlmScoreThresholdLastUpdatedAt :: Lens' MLModel (Maybe UTCTime)

The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

mlmSizeInBytes :: Lens' MLModel (Maybe Integer)

mlmStatus :: Lens' MLModel (Maybe EntityStatus)

The current status of an MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to create an MLModel. INPROGRESS - The creation process is underway. FAILED - The request to create an MLModel did not run to completion. It is not usable. COMPLETED - The creation process completed successfully. DELETED - The MLModel is marked as deleted. It is not usable.

mlmTrainingDataSourceId :: Lens' MLModel (Maybe Text)

The ID of the training DataSource. The CreateMLModel operation uses the TrainingDataSourceId.

mlmTrainingParameters :: Lens' MLModel (HashMap Text Text)

A list of the training parameters in the MLModel. The list is implemented as a map of key/value pairs.

The following is the current set of training parameters:

'sgd.l1RegularizationAmount' - Coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, specify a small value, such as 1.0E-04 or 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is not to use L1 normalization. The parameter cannot be used when L2 is specified. Use this parameter sparingly.

'sgd.l2RegularizationAmount' - Coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, specify a small value, such as 1.0E-04 or 1.0E-08.

The valus is a double that ranges from 0 to MAX_DOUBLE. The default is not to use L2 normalization. This cannot be used when L1 is specified. Use this parameter sparingly.

'sgd.maxPasses' - Number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

'sgd.maxMLModelSizeInBytes' - Maximum allowed size of the model. Depending on the input data, the model size might affect performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

BatchPrediction

data BatchPrediction

batchPrediction :: BatchPrediction

BatchPrediction constructor.

The fields accessible through corresponding lenses are:

bpBatchPredictionDataSourceId :: Maybe Text

bpBatchPredictionId :: Maybe Text

bpCreatedAt :: Maybe UTCTime

bpCreatedByIamUser :: Maybe Text

bpInputDataLocationS3 :: Maybe Text

bpLastUpdatedAt :: Maybe UTCTime

bpMLModelId :: Maybe Text

bpMessage :: Maybe Text

bpName :: Maybe Text

bpOutputUri :: Maybe Text

bpStatus :: Maybe EntityStatus

bpBatchPredictionDataSourceId :: Lens' BatchPrediction (Maybe Text)

The ID of the DataSource that points to the group of observations to predict.

bpBatchPredictionId :: Lens' BatchPrediction (Maybe Text)

The ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

bpCreatedAt :: Lens' BatchPrediction (Maybe UTCTime)

The time that the BatchPrediction was created. The time is expressed in epoch time.

bpCreatedByIamUser :: Lens' BatchPrediction (Maybe Text)

The AWS user account that invoked the BatchPrediction. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

bpInputDataLocationS3 :: Lens' BatchPrediction (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

bpLastUpdatedAt :: Lens' BatchPrediction (Maybe UTCTime)

The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.

bpMLModelId :: Lens' BatchPrediction (Maybe Text)

The ID of the MLModel that generated predictions for the BatchPrediction request.

bpMessage :: Lens' BatchPrediction (Maybe Text)

A description of the most recent details about processing the batch prediction request.

bpName :: Lens' BatchPrediction (Maybe Text)

A user-supplied name or description of the BatchPrediction.

bpOutputUri :: Lens' BatchPrediction (Maybe Text)

The location of an Amazon S3 bucket or directory to receive the operation results. The following substrings are not allowed in the s3 key portion of the "outputURI" field: :, //, /./, /../.

bpStatus :: Lens' BatchPrediction (Maybe EntityStatus)

The status of the BatchPrediction. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate predictions for a batch of observations. INPROGRESS - The process is underway. FAILED - The request to peform a batch prediction did not run to completion. It is not usable. COMPLETED - The batch prediction process completed successfully. DELETED - The BatchPrediction is marked as deleted. It is not usable.

SortOrder

data SortOrder

Constructors

Asc
asc

Dsc
dsc

Algorithm

data Algorithm

Constructors

Sgd
sgd

EntityStatus

data EntityStatus

Constructors

ESCompleted
COMPLETED

ESDeleted
DELETED

ESFailed
FAILED

ESInprogress
INPROGRESS

ESPending
PENDING

DataSource

data DataSource

dataSource :: DataSource

DataSource constructor.

The fields accessible through corresponding lenses are:

dsComputeStatistics :: Maybe Bool

dsCreatedAt :: Maybe UTCTime

dsCreatedByIamUser :: Maybe Text

dsDataLocationS3 :: Maybe Text

dsDataRearrangement :: Maybe Text

dsDataSizeInBytes :: Maybe Integer

dsDataSourceId :: Maybe Text

dsLastUpdatedAt :: Maybe UTCTime

dsMessage :: Maybe Text

dsName :: Maybe Text

dsNumberOfFiles :: Maybe Integer

dsRDSMetadata :: Maybe RDSMetadata

dsRedshiftMetadata :: Maybe RedshiftMetadata

dsRoleARN :: Maybe Text

dsStatus :: Maybe EntityStatus

dsComputeStatistics :: Lens' DataSource (Maybe Bool)

The parameter is true if statistics need to be generated from the observation data.

dsCreatedAt :: Lens' DataSource (Maybe UTCTime)

The time that the DataSource was created. The time is expressed in epoch time.

dsCreatedByIamUser :: Lens' DataSource (Maybe Text)

The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

dsDataLocationS3 :: Lens' DataSource (Maybe Text)

The location and name of the data in Amazon Simple Storage Service (Amazon S3) that is used by a DataSource.

dsDataRearrangement :: Lens' DataSource (Maybe Text)

A JSON string that represents the splitting requirement of a Datasource.

dsDataSizeInBytes :: Lens' DataSource (Maybe Integer)

The total number of observations contained in the data files that the DataSource references.

dsDataSourceId :: Lens' DataSource (Maybe Text)

The ID that is assigned to the DataSource during creation.

dsLastUpdatedAt :: Lens' DataSource (Maybe UTCTime)

The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.

dsMessage :: Lens' DataSource (Maybe Text)

A description of the most recent details about creating the DataSource.

dsName :: Lens' DataSource (Maybe Text)

A user-supplied name or description of the DataSource.

dsNumberOfFiles :: Lens' DataSource (Maybe Integer)

The number of data files referenced by the DataSource.

dsRDSMetadata :: Lens' DataSource (Maybe RDSMetadata)

dsRedshiftMetadata :: Lens' DataSource (Maybe RedshiftMetadata)

dsRoleARN :: Lens' DataSource (Maybe Text)

dsStatus :: Lens' DataSource (Maybe EntityStatus)

The current status of the DataSource. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to create a DataSource. INPROGRESS - The creation process is underway. FAILED - The request to create a DataSource did not run to completion. It is not usable. COMPLETED - The creation process completed successfully. DELETED - The DataSource is marked as deleted. It is not usable.

RDSDatabase

data RDSDatabase

rdsdatabase

Arguments

:: Text
rdsdInstanceIdentifier

-> Text
rdsdDatabaseName

-> RDSDatabase

RDSDatabase constructor.

The fields accessible through corresponding lenses are:

rdsdDatabaseName :: Text

rdsdInstanceIdentifier :: Text

rdsdDatabaseName :: Lens' RDSDatabase Text

rdsdInstanceIdentifier :: Lens' RDSDatabase Text

The ID of an RDS DB instance.

RDSDatabaseCredentials

data RDSDatabaseCredentials

rdsdatabaseCredentials

Arguments

:: Text
rdsdcUsername

-> Text
rdsdcPassword

-> RDSDatabaseCredentials

RDSDatabaseCredentials constructor.

The fields accessible through corresponding lenses are:

rdsdcPassword :: Text

rdsdcUsername :: Text

rdsdcPassword :: Lens' RDSDatabaseCredentials Text

rdsdcUsername :: Lens' RDSDatabaseCredentials Text

MLModelFilterVariable

data MLModelFilterVariable

Constructors

MLMFVAlgorithm
Algorithm

MLMFVCreatedAt
CreatedAt

MLMFVIAMUser
IAMUser

MLMFVLastUpdatedAt
LastUpdatedAt

MLMFVMLModelType
MLModelType

MLMFVName
Name

MLMFVRealtimeEndpointStatus
RealtimeEndpointStatus

MLMFVStatus
Status

MLMFVTrainingDataSourceId
TrainingDataSourceId

MLMFVTrainingDataURI
TrainingDataURI

DataSourceFilterVariable

data DataSourceFilterVariable

Constructors

DataCreatedAt
CreatedAt

DataDataLocationS3
DataLocationS3

DataIAMUser
IAMUser

DataLastUpdatedAt
LastUpdatedAt

DataName
Name

DataStatus
Status

RedshiftDataSpec

data RedshiftDataSpec

redshiftDataSpec

Arguments

:: RedshiftDatabase
rdsDatabaseInformation

-> Text
rdsSelectSqlQuery

-> RedshiftDatabaseCredentials
rdsDatabaseCredentials

-> Text
rdsS3StagingLocation

-> RedshiftDataSpec

RedshiftDataSpec constructor.

The fields accessible through corresponding lenses are:

rdsDataRearrangement :: Maybe Text

rdsDataSchema :: Maybe Text

rdsDataSchemaUri :: Maybe Text

rdsDatabaseCredentials :: RedshiftDatabaseCredentials

rdsDatabaseInformation :: RedshiftDatabase

rdsS3StagingLocation :: Text

rdsSelectSqlQuery :: Text

rdsDataRearrangement :: Lens' RedshiftDataSpec (Maybe Text)

Describes the splitting specifications for a DataSource.

rdsDataSchema :: Lens' RedshiftDataSpec (Maybe Text)

Describes the schema for an Amazon Redshift DataSource.

rdsDataSchemaUri :: Lens' RedshiftDataSpec (Maybe Text)

Describes the schema location for an Amazon Redshift DataSource.

rdsDatabaseCredentials :: Lens' RedshiftDataSpec RedshiftDatabaseCredentials

Describes AWS Identity and Access Management (IAM) credentials that are used connect to the Amazon Redshift database.

rdsDatabaseInformation :: Lens' RedshiftDataSpec RedshiftDatabase

Describes the DatabaseName and ClusterIdentifier for an Amazon Redshift DataSource.

rdsS3StagingLocation :: Lens' RedshiftDataSpec Text

Describes an Amazon S3 location to store the result set of the SelectSqlQuery query.

rdsSelectSqlQuery :: Lens' RedshiftDataSpec Text

Describes the SQL Query to execute on an Amazon Redshift database for an Amazon Redshift DataSource.

Evaluation

data Evaluation

evaluation :: Evaluation

Evaluation constructor.

The fields accessible through corresponding lenses are:

eCreatedAt :: Maybe UTCTime

eCreatedByIamUser :: Maybe Text

eEvaluationDataSourceId :: Maybe Text

eEvaluationId :: Maybe Text

eInputDataLocationS3 :: Maybe Text

eLastUpdatedAt :: Maybe UTCTime

eMLModelId :: Maybe Text

eMessage :: Maybe Text

eName :: Maybe Text

ePerformanceMetrics :: Maybe PerformanceMetrics

eStatus :: Maybe EntityStatus

eCreatedAt :: Lens' Evaluation (Maybe UTCTime)

The time that the Evaluation was created. The time is expressed in epoch time.

eCreatedByIamUser :: Lens' Evaluation (Maybe Text)

The AWS user account that invoked the evaluation. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

eEvaluationDataSourceId :: Lens' Evaluation (Maybe Text)

The ID of the DataSource that is used to evaluate the MLModel.

eEvaluationId :: Lens' Evaluation (Maybe Text)

The ID that is assigned to the Evaluation at creation.

eInputDataLocationS3 :: Lens' Evaluation (Maybe Text)

The location and name of the data in Amazon Simple Storage Server (Amazon S3) that is used in the evaluation.

eLastUpdatedAt :: Lens' Evaluation (Maybe UTCTime)

The time of the most recent edit to the Evaluation. The time is expressed in epoch time.

eMLModelId :: Lens' Evaluation (Maybe Text)

The ID of the MLModel that is the focus of the evaluation.

eMessage :: Lens' Evaluation (Maybe Text)

A description of the most recent details about evaluating the MLModel.

eName :: Lens' Evaluation (Maybe Text)

A user-supplied name or description of the Evaluation.

ePerformanceMetrics :: Lens' Evaluation (Maybe PerformanceMetrics)

Measurements of how well the MLModel performed, using observations referenced by the DataSource. One of the following metrics is returned, based on the type of the MLModel:

BinaryAUC: A binary MLModel uses the Area Under the Curve (AUC) technique to measure performance.

RegressionRMSE: A regression MLModel uses the Root Mean Square Error (RMSE) technique to measure performance. RMSE measures the difference between predicted and actual values for a single variable.

MulticlassAvgFScore: A multiclass MLModel uses the F1 score technique to measure performance.

For more information about performance metrics, please see the AmazonMachine Learning Developer Guide.

eStatus :: Lens' Evaluation (Maybe EntityStatus)

The status of the evaluation. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to evaluate an MLModel. INPROGRESS - The evaluation is underway. FAILED - The request to evaluate an MLModel did not run to completion. It is not usable. COMPLETED - The evaluation process completed successfully. DELETED - The Evaluation is marked as deleted. It is not usable.

RedshiftMetadata

data RedshiftMetadata

redshiftMetadata :: RedshiftMetadata

RedshiftMetadata constructor.

The fields accessible through corresponding lenses are:

rmDatabaseUserName :: Maybe Text

rmRedshiftDatabase :: Maybe RedshiftDatabase

rmSelectSqlQuery :: Maybe Text

rmDatabaseUserName :: Lens' RedshiftMetadata (Maybe Text)

rmRedshiftDatabase :: Lens' RedshiftMetadata (Maybe RedshiftDatabase)

rmSelectSqlQuery :: Lens' RedshiftMetadata (Maybe Text)

The SQL query that is specified during CreateDataSourceFromRedshift. Returns only if Verbose is true in GetDataSourceInput.

MLModelType

data MLModelType

Constructors

Binary
BINARY

Multiclass
MULTICLASS

Regression
REGRESSION