Network.AWS.MachineLearning.UpdateBatchPrediction

Description

Updates the BatchPredictionName of a BatchPrediction.

You can use the GetBatchPrediction operation to view the contents of the updated data element.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_UpdateBatchPrediction.html

Request

data UpdateBatchPrediction

Request constructor

updateBatchPrediction

Arguments

:: Text
ubpBatchPredictionId

-> Text
ubpBatchPredictionName

-> UpdateBatchPrediction

UpdateBatchPrediction constructor.

The fields accessible through corresponding lenses are:

ubpBatchPredictionId :: Text

ubpBatchPredictionName :: Text

Request lenses

ubpBatchPredictionId :: Lens' UpdateBatchPrediction Text

The ID assigned to the BatchPrediction during creation.

ubpBatchPredictionName :: Lens' UpdateBatchPrediction Text

A new user-supplied name or description of the BatchPrediction.

Response

data UpdateBatchPredictionResponse

Response constructor

updateBatchPredictionResponse :: UpdateBatchPredictionResponse

UpdateBatchPredictionResponse constructor.

The fields accessible through corresponding lenses are:

ubprBatchPredictionId :: Maybe Text

Response lenses

ubprBatchPredictionId :: Lens' UpdateBatchPredictionResponse (Maybe Text)

The ID assigned to the BatchPrediction during creation. This value should be identical to the value of the BatchPredictionId in the request.