Network.AWS.MachineLearning.UpdateDataSource

Description

Updates the DataSourceName of a DataSource.

You can use the GetDataSource operation to view the contents of the updated data element.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_UpdateDataSource.html

Request

data UpdateDataSource

Request constructor

updateDataSource

Arguments

:: Text
udsDataSourceId

-> Text
udsDataSourceName

-> UpdateDataSource

UpdateDataSource constructor.

The fields accessible through corresponding lenses are:

udsDataSourceId :: Text

udsDataSourceName :: Text

Request lenses

udsDataSourceId :: Lens' UpdateDataSource Text

The ID assigned to the DataSource during creation.

udsDataSourceName :: Lens' UpdateDataSource Text

A new user-supplied name or description of the DataSource that will replace the current description.

Response

data UpdateDataSourceResponse

Response constructor

updateDataSourceResponse :: UpdateDataSourceResponse

UpdateDataSourceResponse constructor.

The fields accessible through corresponding lenses are:

udsrDataSourceId :: Maybe Text

Response lenses

udsrDataSourceId :: Lens' UpdateDataSourceResponse (Maybe Text)

The ID assigned to the DataSource during creation. This value should be identical to the value of the DataSourceID in the request.