Network.AWS.MachineLearning.UpdateEvaluation

Description

Updates the EvaluationName of an Evaluation.

You can use the GetEvaluation operation to view the contents of the updated data element.

http://http://docs.aws.amazon.com/machine-learning/latest/APIReference/API_UpdateEvaluation.html

Request

data UpdateEvaluation

Request constructor

updateEvaluation

Arguments

:: Text
ueEvaluationId

-> Text
ueEvaluationName

-> UpdateEvaluation

UpdateEvaluation constructor.

The fields accessible through corresponding lenses are:

ueEvaluationId :: Text

ueEvaluationName :: Text

Request lenses

ueEvaluationId :: Lens' UpdateEvaluation Text

The ID assigned to the Evaluation during creation.

ueEvaluationName :: Lens' UpdateEvaluation Text

A new user-supplied name or description of the Evaluation that will replace the current content.

Response

data UpdateEvaluationResponse

Response constructor

updateEvaluationResponse :: UpdateEvaluationResponse

UpdateEvaluationResponse constructor.

The fields accessible through corresponding lenses are:

uerEvaluationId :: Maybe Text

Response lenses

uerEvaluationId :: Lens' UpdateEvaluationResponse (Maybe Text)

The ID assigned to the Evaluation during creation. This value should be identical to the value of the Evaluation in the request.