Network.AWS.MachineLearning.UpdateMLModel

Description

Updates the MLModelName and the ScoreThreshold of an MLModel .

You can use the GetMLModel operation to view the contents of the updated data element.

Creating a Request

updateMLModel

Arguments

:: Text
umlmMLModelId

-> UpdateMLModel

Creates a value of UpdateMLModel with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umlmMLModelName - A user-supplied name or description of the MLModel .

umlmScoreThreshold - The ScoreThreshold used in binary classification MLModel that marks the boundary between a positive prediction and a negative prediction. Output values greater than or equal to the ScoreThreshold receive a positive result from the MLModel , such as true . Output values less than the ScoreThreshold receive a negative response from the MLModel , such as false .

umlmMLModelId - The ID assigned to the MLModel during creation.

data UpdateMLModel

See: updateMLModel smart constructor.

Request Lenses

umlmMLModelName :: Lens' UpdateMLModel (Maybe Text)

A user-supplied name or description of the MLModel .

umlmScoreThreshold :: Lens' UpdateMLModel (Maybe Double)

The ScoreThreshold used in binary classification MLModel that marks the boundary between a positive prediction and a negative prediction. Output values greater than or equal to the ScoreThreshold receive a positive result from the MLModel , such as true . Output values less than the ScoreThreshold receive a negative response from the MLModel , such as false .

umlmMLModelId :: Lens' UpdateMLModel Text

The ID assigned to the MLModel during creation.

Destructuring the Response

updateMLModelResponse

Arguments

:: Int
umlmrsResponseStatus

-> UpdateMLModelResponse

Creates a value of UpdateMLModelResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umlmrsMLModelId - The ID assigned to the MLModel during creation. This value should be identical to the value of the MLModelID in the request.

umlmrsResponseStatus - -- | The response status code.

data UpdateMLModelResponse

Represents the output of an UpdateMLModel operation.

You can see the updated content by using the GetMLModel operation.

See: updateMLModelResponse smart constructor.

Response Lenses

umlmrsMLModelId :: Lens' UpdateMLModelResponse (Maybe Text)

The ID assigned to the MLModel during creation. This value should be identical to the value of the MLModelID in the request.

umlmrsResponseStatus :: Lens' UpdateMLModelResponse Int

- | The response status code.