Network.AWS.MachineLearning.CreateBatchPrediction

Description

Generates predictions for a group of observations. The observations to process exist in one or more data files referenced by a DataSource . This operation creates a new BatchPrediction , and uses an MLModel and the data files referenced by the DataSource as information sources.

CreateBatchPrediction is an asynchronous operation. In response to CreateBatchPrediction , Amazon Machine Learning (Amazon ML) immediately returns and sets the BatchPrediction status to PENDING . After the BatchPrediction completes, Amazon ML sets the status to COMPLETED .

You can poll for status updates by using the GetBatchPrediction operation and checking the Status parameter of the result. After the COMPLETED status appears, the results are available in the location specified by the OutputUri parameter.

Creating a Request

createBatchPrediction

Arguments

:: Text
cbpBatchPredictionId

-> Text
cbpMLModelId

-> Text
cbpBatchPredictionDataSourceId

-> Text
cbpOutputURI

-> CreateBatchPrediction

Creates a value of CreateBatchPrediction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbpBatchPredictionName - A user-supplied name or description of the BatchPrediction . BatchPredictionName can only use the UTF-8 character set.

cbpBatchPredictionId - A user-supplied ID that uniquely identifies the BatchPrediction .

cbpMLModelId - The ID of the MLModel that will generate predictions for the group of observations.

cbpBatchPredictionDataSourceId - The ID of the DataSource that points to the group of observations to predict.

cbpOutputURI - The location of an Amazon Simple Storage Service (Amazon S3) bucket or directory to store the batch prediction results. The following substrings are not allowed in the s3 key portion of the outputURI field: :, //, /./, /../. Amazon ML needs permissions to store and retrieve the logs on your behalf. For information about how to set permissions, see the Amazon Machine Learning Developer Guide .

data CreateBatchPrediction

See: createBatchPrediction smart constructor.

Request Lenses

cbpBatchPredictionName :: Lens' CreateBatchPrediction (Maybe Text)

A user-supplied name or description of the BatchPrediction . BatchPredictionName can only use the UTF-8 character set.

cbpBatchPredictionId :: Lens' CreateBatchPrediction Text

A user-supplied ID that uniquely identifies the BatchPrediction .

cbpMLModelId :: Lens' CreateBatchPrediction Text

The ID of the MLModel that will generate predictions for the group of observations.

cbpBatchPredictionDataSourceId :: Lens' CreateBatchPrediction Text

The ID of the DataSource that points to the group of observations to predict.

cbpOutputURI :: Lens' CreateBatchPrediction Text

The location of an Amazon Simple Storage Service (Amazon S3) bucket or directory to store the batch prediction results. The following substrings are not allowed in the s3 key portion of the outputURI field: :, //, /./, /../. Amazon ML needs permissions to store and retrieve the logs on your behalf. For information about how to set permissions, see the Amazon Machine Learning Developer Guide .

Destructuring the Response

createBatchPredictionResponse

Arguments

:: Int
cbprsResponseStatus

-> CreateBatchPredictionResponse

Creates a value of CreateBatchPredictionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbprsBatchPredictionId - A user-supplied ID that uniquely identifies the BatchPrediction . This value is identical to the value of the BatchPredictionId in the request.

cbprsResponseStatus - -- | The response status code.

data CreateBatchPredictionResponse

Represents the output of a CreateBatchPrediction operation, and is an acknowledgement that Amazon ML received the request.

The CreateBatchPrediction operation is asynchronous. You can poll for status updates by using the >GetBatchPrediction operation and checking the Status parameter of the result.

See: createBatchPredictionResponse smart constructor.

Response Lenses

cbprsBatchPredictionId :: Lens' CreateBatchPredictionResponse (Maybe Text)

A user-supplied ID that uniquely identifies the BatchPrediction . This value is identical to the value of the BatchPredictionId in the request.

cbprsResponseStatus :: Lens' CreateBatchPredictionResponse Int

- | The response status code.