Network.AWS.MachineLearning.CreateDataSourceFromRDS

Description

Creates a DataSource object from an Amazon Relational Database Service (Amazon RDS). A DataSource references data that can be used to perform CreateMLModel , CreateEvaluation , or CreateBatchPrediction operations.

CreateDataSourceFromRDS is an asynchronous operation. In response to CreateDataSourceFromRDS , Amazon Machine Learning (Amazon ML) immediately returns and sets the DataSource status to PENDING . After the DataSource is created and ready for use, Amazon ML sets the Status parameter to COMPLETED . DataSource in the COMPLETED or PENDING state can be used only to perform >CreateMLModel >, CreateEvaluation , or CreateBatchPrediction operations.

If Amazon ML cannot accept the input source, it sets the Status parameter to FAILED and includes an error message in the Message attribute of the GetDataSource operation response.

Creating a Request

createDataSourceFromRDS

Arguments

:: Text
cdsfrdsDataSourceId

-> RDSDataSpec
cdsfrdsRDSData

-> Text
cdsfrdsRoleARN

-> CreateDataSourceFromRDS

Creates a value of CreateDataSourceFromRDS with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdsfrdsDataSourceName - A user-supplied name or description of the DataSource .

cdsfrdsComputeStatistics - The compute statistics for a DataSource . The statistics are generated from the observation data referenced by a DataSource . Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the DataSourceneeds to be used for MLModel training.

cdsfrdsDataSourceId - A user-supplied ID that uniquely identifies the DataSource . Typically, an Amazon Resource Number (ARN) becomes the ID for a DataSource .

cdsfrdsRDSData - The data specification of an Amazon RDS DataSource : * DatabaseInformation - * DatabaseName - The name of the Amazon RDS database. * InstanceIdentifier - A unique identifier for the Amazon RDS database instance. * DatabaseCredentials - AWS Identity and Access Management (IAM) credentials that are used to connect to the Amazon RDS database. * ResourceRole - A role (DataPipelineDefaultResourceRole) assumed by an EC2 instance to carry out the copy task from Amazon RDS to Amazon Simple Storage Service (Amazon S3). For more information, see Role templates for data pipelines. * ServiceRole - A role (DataPipelineDefaultRole) assumed by the AWS Data Pipeline service to monitor the progress of the copy task from Amazon RDS to Amazon S3. For more information, see Role templates for data pipelines. * SecurityInfo - The security information to use to access an RDS DB instance. You need to set up appropriate ingress rules for the security entity IDs provided to allow access to the Amazon RDS instance. Specify a [SubnetId , SecurityGroupIds ] pair for a VPC-based RDS DB instance. * SelectSqlQuery - A query that is used to retrieve the observation data for the Datasource . * S3StagingLocation - The Amazon S3 location for staging Amazon RDS data. The data retrieved from Amazon RDS using SelectSqlQuery is stored in this location. * DataSchemaUri - The Amazon S3 location of the DataSchema . * DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified. * DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the Datasource . Sample - "{"splitting":{"percentBegin":10,"percentEnd":60}}"

cdsfrdsRoleARN - The role that Amazon ML assumes on behalf of the user to create and activate a data pipeline in the user's account and copy data using the SelectSqlQuery query from Amazon RDS to Amazon S3.

data CreateDataSourceFromRDS

See: createDataSourceFromRDS smart constructor.

Request Lenses

cdsfrdsDataSourceName :: Lens' CreateDataSourceFromRDS (Maybe Text)

A user-supplied name or description of the DataSource .

cdsfrdsComputeStatistics :: Lens' CreateDataSourceFromRDS (Maybe Bool)

The compute statistics for a DataSource . The statistics are generated from the observation data referenced by a DataSource . Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the DataSourceneeds to be used for MLModel training.

cdsfrdsDataSourceId :: Lens' CreateDataSourceFromRDS Text

A user-supplied ID that uniquely identifies the DataSource . Typically, an Amazon Resource Number (ARN) becomes the ID for a DataSource .

cdsfrdsRDSData :: Lens' CreateDataSourceFromRDS RDSDataSpec

The data specification of an Amazon RDS DataSource : * DatabaseInformation - * DatabaseName - The name of the Amazon RDS database. * InstanceIdentifier - A unique identifier for the Amazon RDS database instance. * DatabaseCredentials - AWS Identity and Access Management (IAM) credentials that are used to connect to the Amazon RDS database. * ResourceRole - A role (DataPipelineDefaultResourceRole) assumed by an EC2 instance to carry out the copy task from Amazon RDS to Amazon Simple Storage Service (Amazon S3). For more information, see Role templates for data pipelines. * ServiceRole - A role (DataPipelineDefaultRole) assumed by the AWS Data Pipeline service to monitor the progress of the copy task from Amazon RDS to Amazon S3. For more information, see Role templates for data pipelines. * SecurityInfo - The security information to use to access an RDS DB instance. You need to set up appropriate ingress rules for the security entity IDs provided to allow access to the Amazon RDS instance. Specify a [SubnetId , SecurityGroupIds ] pair for a VPC-based RDS DB instance. * SelectSqlQuery - A query that is used to retrieve the observation data for the Datasource . * S3StagingLocation - The Amazon S3 location for staging Amazon RDS data. The data retrieved from Amazon RDS using SelectSqlQuery is stored in this location. * DataSchemaUri - The Amazon S3 location of the DataSchema . * DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified. * DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the Datasource . Sample - "{"splitting":{"percentBegin":10,"percentEnd":60}}"

cdsfrdsRoleARN :: Lens' CreateDataSourceFromRDS Text

The role that Amazon ML assumes on behalf of the user to create and activate a data pipeline in the user's account and copy data using the SelectSqlQuery query from Amazon RDS to Amazon S3.

Destructuring the Response

createDataSourceFromRDSResponse

Arguments

:: Int
cdsfrdsrsResponseStatus

-> CreateDataSourceFromRDSResponse

Creates a value of CreateDataSourceFromRDSResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdsfrdsrsDataSourceId - A user-supplied ID that uniquely identifies the datasource. This value should be identical to the value of the DataSourceID in the request.

cdsfrdsrsResponseStatus - -- | The response status code.

data CreateDataSourceFromRDSResponse

Represents the output of a CreateDataSourceFromRDS operation, and is an acknowledgement that Amazon ML received the request.

The CreateDataSourceFromRDS > operation is asynchronous. You can poll for updates by using the GetBatchPrediction operation and checking the Status parameter. You can inspect the Message when Status shows up as FAILED . You can also check the progress of the copy operation by going to the DataPipeline console and looking up the pipeline using the pipelineId from the describe call.

See: createDataSourceFromRDSResponse smart constructor.

Response Lenses

cdsfrdsrsDataSourceId :: Lens' CreateDataSourceFromRDSResponse (Maybe Text)

A user-supplied ID that uniquely identifies the datasource. This value should be identical to the value of the DataSourceID in the request.

cdsfrdsrsResponseStatus :: Lens' CreateDataSourceFromRDSResponse Int

- | The response status code.