Network.AWS.MachineLearning.CreateDataSourceFromRedshift

Description

Creates a DataSource from a database hosted on an Amazon Redshift cluster. A DataSource references data that can be used to perform either CreateMLModel , CreateEvaluation , or CreateBatchPrediction operations.

CreateDataSourceFromRedshift is an asynchronous operation. In response to CreateDataSourceFromRedshift , Amazon Machine Learning (Amazon ML) immediately returns and sets the DataSource status to PENDING . After the DataSource is created and ready for use, Amazon ML sets the Status parameter to COMPLETED . DataSource in COMPLETED or PENDING states can be used to perform only CreateMLModel , CreateEvaluation , or CreateBatchPrediction operations.

If Amazon ML can't accept the input source, it sets the Status parameter to FAILED and includes an error message in the Message attribute of the GetDataSource operation response.

The observations should be contained in the database hosted on an Amazon Redshift cluster and should be specified by a SelectSqlQuery query. Amazon ML executes an Unload command in Amazon Redshift to transfer the result set of the SelectSqlQuery query to S3StagingLocation .

After the DataSource has been created, it's ready for use in evaluations and batch predictions. If you plan to use the DataSource to train an MLModel , the DataSource also requires a recipe. A recipe describes how each input variable will be used in training an MLModel . Will the variable be included or excluded from training? Will the variable be manipulated; for example, will it be combined with another variable or will it be split apart into word combinations? The recipe provides answers to these questions.

You can't change an existing datasource, but you can copy and modify the settings from an existing Amazon Redshift datasource to create a new datasource. To do so, call GetDataSource for an existing datasource and copy the values to a CreateDataSource call. Change the settings that you want to change and make sure that all required fields have the appropriate values.

Creating a Request

createDataSourceFromRedshift

Arguments

:: Text
cdsfrDataSourceId

-> RedshiftDataSpec
cdsfrDataSpec

-> Text
cdsfrRoleARN

-> CreateDataSourceFromRedshift

Creates a value of CreateDataSourceFromRedshift with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdsfrDataSourceName - A user-supplied name or description of the DataSource .

cdsfrComputeStatistics - The compute statistics for a DataSource . The statistics are generated from the observation data referenced by a DataSource . Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the DataSource needs to be used for MLModel training.

cdsfrDataSourceId - A user-supplied ID that uniquely identifies the DataSource .

cdsfrDataSpec - The data specification of an Amazon Redshift DataSource : * DatabaseInformation - * DatabaseName - The name of the Amazon Redshift database. * ClusterIdentifier - The unique ID for the Amazon Redshift cluster. * DatabaseCredentials - The AWS Identity and Access Management (IAM) credentials that are used to connect to the Amazon Redshift database. * SelectSqlQuery - The query that is used to retrieve the observation data for the Datasource . * S3StagingLocation - The Amazon Simple Storage Service (Amazon S3) location for staging Amazon Redshift data. The data retrieved from Amazon Redshift using the SelectSqlQuery query is stored in this location. * DataSchemaUri - The Amazon S3 location of the DataSchema . * DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified. * DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the DataSource . Sample - "{"splitting":{"percentBegin":10,"percentEnd":60}}"

cdsfrRoleARN - A fully specified role Amazon Resource Name (ARN). Amazon ML assumes the role on behalf of the user to create the following: * A security group to allow Amazon ML to execute the SelectSqlQuery query on an Amazon Redshift cluster * An Amazon S3 bucket policy to grant Amazon ML read/write permissions on the S3StagingLocation

data CreateDataSourceFromRedshift

See: createDataSourceFromRedshift smart constructor.

Request Lenses

cdsfrDataSourceName :: Lens' CreateDataSourceFromRedshift (Maybe Text)

A user-supplied name or description of the DataSource .

cdsfrComputeStatistics :: Lens' CreateDataSourceFromRedshift (Maybe Bool)

The compute statistics for a DataSource . The statistics are generated from the observation data referenced by a DataSource . Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the DataSource needs to be used for MLModel training.

cdsfrDataSourceId :: Lens' CreateDataSourceFromRedshift Text

A user-supplied ID that uniquely identifies the DataSource .

cdsfrDataSpec :: Lens' CreateDataSourceFromRedshift RedshiftDataSpec

The data specification of an Amazon Redshift DataSource : * DatabaseInformation - * DatabaseName - The name of the Amazon Redshift database. * ClusterIdentifier - The unique ID for the Amazon Redshift cluster. * DatabaseCredentials - The AWS Identity and Access Management (IAM) credentials that are used to connect to the Amazon Redshift database. * SelectSqlQuery - The query that is used to retrieve the observation data for the Datasource . * S3StagingLocation - The Amazon Simple Storage Service (Amazon S3) location for staging Amazon Redshift data. The data retrieved from Amazon Redshift using the SelectSqlQuery query is stored in this location. * DataSchemaUri - The Amazon S3 location of the DataSchema . * DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified. * DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the DataSource . Sample - "{"splitting":{"percentBegin":10,"percentEnd":60}}"

cdsfrRoleARN :: Lens' CreateDataSourceFromRedshift Text

A fully specified role Amazon Resource Name (ARN). Amazon ML assumes the role on behalf of the user to create the following: * A security group to allow Amazon ML to execute the SelectSqlQuery query on an Amazon Redshift cluster * An Amazon S3 bucket policy to grant Amazon ML read/write permissions on the S3StagingLocation

Destructuring the Response

createDataSourceFromRedshiftResponse

Arguments

:: Int
cdsfrrsResponseStatus

-> CreateDataSourceFromRedshiftResponse

Creates a value of CreateDataSourceFromRedshiftResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdsfrrsDataSourceId - A user-supplied ID that uniquely identifies the datasource. This value should be identical to the value of the DataSourceID in the request.

cdsfrrsResponseStatus - -- | The response status code.

data CreateDataSourceFromRedshiftResponse

Represents the output of a CreateDataSourceFromRedshift operation, and is an acknowledgement that Amazon ML received the request.

The CreateDataSourceFromRedshift operation is asynchronous. You can poll for updates by using the GetBatchPrediction operation and checking the Status parameter.

See: createDataSourceFromRedshiftResponse smart constructor.

Response Lenses

cdsfrrsDataSourceId :: Lens' CreateDataSourceFromRedshiftResponse (Maybe Text)

A user-supplied ID that uniquely identifies the datasource. This value should be identical to the value of the DataSourceID in the request.

cdsfrrsResponseStatus :: Lens' CreateDataSourceFromRedshiftResponse Int

- | The response status code.