Network.AWS.MachineLearning.CreateEvaluation

Description

Creates a new Evaluation of an MLModel . An MLModel is evaluated on a set of observations associated to a DataSource . Like a DataSource for an MLModel , the DataSource for an Evaluation contains values for the Target Variable . The Evaluation compares the predicted result for each observation to the actual outcome and provides a summary so that you know how effective the MLModel functions on the test data. Evaluation generates a relevant performance metric, such as BinaryAUC, RegressionRMSE or MulticlassAvgFScore based on the corresponding MLModelType : BINARY , REGRESSION or MULTICLASS .

CreateEvaluation is an asynchronous operation. In response to CreateEvaluation , Amazon Machine Learning (Amazon ML) immediately returns and sets the evaluation status to PENDING . After the Evaluation is created and ready for use, Amazon ML sets the status to COMPLETED .

You can use the GetEvaluation operation to check progress of the evaluation during the creation operation.

Creating a Request

createEvaluation

Arguments

:: Text
ceEvaluationId

-> Text
ceMLModelId

-> Text
ceEvaluationDataSourceId

-> CreateEvaluation

Creates a value of CreateEvaluation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ceEvaluationName - A user-supplied name or description of the Evaluation .

ceEvaluationId - A user-supplied ID that uniquely identifies the Evaluation .

ceMLModelId - The ID of the MLModel to evaluate. The schema used in creating the MLModel must match the schema of the DataSource used in the Evaluation .

ceEvaluationDataSourceId - The ID of the DataSource for the evaluation. The schema of the DataSource must match the schema used to create the MLModel .

data CreateEvaluation

See: createEvaluation smart constructor.

Request Lenses

ceEvaluationName :: Lens' CreateEvaluation (Maybe Text)

A user-supplied name or description of the Evaluation .

ceEvaluationId :: Lens' CreateEvaluation Text

A user-supplied ID that uniquely identifies the Evaluation .

ceMLModelId :: Lens' CreateEvaluation Text

The ID of the MLModel to evaluate. The schema used in creating the MLModel must match the schema of the DataSource used in the Evaluation .

ceEvaluationDataSourceId :: Lens' CreateEvaluation Text

The ID of the DataSource for the evaluation. The schema of the DataSource must match the schema used to create the MLModel .

Destructuring the Response

createEvaluationResponse

Arguments

:: Int
cersResponseStatus

-> CreateEvaluationResponse

Creates a value of CreateEvaluationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cersEvaluationId - The user-supplied ID that uniquely identifies the Evaluation . This value should be identical to the value of the EvaluationId in the request.

cersResponseStatus - -- | The response status code.

data CreateEvaluationResponse

Represents the output of a CreateEvaluation operation, and is an acknowledgement that Amazon ML received the request.

CreateEvaluation operation is asynchronous. You can poll for status updates by using the GetEvcaluation operation and checking the Status parameter.

See: createEvaluationResponse smart constructor.

Response Lenses

cersEvaluationId :: Lens' CreateEvaluationResponse (Maybe Text)

The user-supplied ID that uniquely identifies the Evaluation . This value should be identical to the value of the EvaluationId in the request.

cersResponseStatus :: Lens' CreateEvaluationResponse Int

- | The response status code.