Network.AWS.MachineLearning.DescribeBatchPredictions

Description

Returns a list of BatchPrediction operations that match the search criteria in the request.

This operation returns paginated results.

Creating a Request

describeBatchPredictions :: DescribeBatchPredictions

Creates a value of DescribeBatchPredictions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbpEQ - The equal to operator. The BatchPrediction results will have FilterVariable values that exactly match the value specified with EQ .

dbpGE - The greater than or equal to operator. The BatchPrediction results will have FilterVariable values that are greater than or equal to the value specified with GE .

dbpPrefix - A string that is found at the beginning of a variable, such as Name or Id . For example, a Batch Prediction operation could have the Name 2014-09-09-HolidayGiftMailer . To search for this BatchPrediction , select Name for the FilterVariable and any of the following strings for the Prefix : * 2014-09 * 2014-09-09 * 2014-09-09-Holiday

dbpGT - The greater than operator. The BatchPrediction results will have FilterVariable values that are greater than the value specified with GT .

dbpNE - The not equal to operator. The BatchPrediction results will have FilterVariable values not equal to the value specified with NE .

dbpNextToken - An ID of the page in the paginated results.

dbpSortOrder - A two-value parameter that determines the sequence of the resulting list of MLModel s. * asc - Arranges the list in ascending order (A-Z, 0-9). * dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable .

dbpLimit - The number of pages of information to include in the result. The range of acceptable values is 1 through 100 . The default value is 100 .

dbpLT - The less than operator. The BatchPrediction results will have FilterVariable values that are less than the value specified with LT .

dbpFilterVariable - Use one of the following variables to filter a list of BatchPrediction : * CreatedAt - Sets the search criteria to the BatchPrediction creation date. * Status - Sets the search criteria to the BatchPrediction status. * Name - Sets the search criteria to the contents of the BatchPrediction ____ Name . * IAMUser - Sets the search criteria to the user account that invoked the BatchPrediction creation. * MLModelId - Sets the search criteria to the MLModel used in the BatchPrediction . * DataSourceId - Sets the search criteria to the DataSource used in the BatchPrediction . * DataURI - Sets the search criteria to the data file(s) used in the BatchPrediction . The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

dbpLE - The less than or equal to operator. The BatchPrediction results will have FilterVariable values that are less than or equal to the value specified with LE .

data DescribeBatchPredictions

See: describeBatchPredictions smart constructor.

Request Lenses

dbpEQ :: Lens' DescribeBatchPredictions (Maybe Text)

The equal to operator. The BatchPrediction results will have FilterVariable values that exactly match the value specified with EQ .

dbpGE :: Lens' DescribeBatchPredictions (Maybe Text)

The greater than or equal to operator. The BatchPrediction results will have FilterVariable values that are greater than or equal to the value specified with GE .

dbpPrefix :: Lens' DescribeBatchPredictions (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id . For example, a Batch Prediction operation could have the Name 2014-09-09-HolidayGiftMailer . To search for this BatchPrediction , select Name for the FilterVariable and any of the following strings for the Prefix : * 2014-09 * 2014-09-09 * 2014-09-09-Holiday

dbpGT :: Lens' DescribeBatchPredictions (Maybe Text)

The greater than operator. The BatchPrediction results will have FilterVariable values that are greater than the value specified with GT .

dbpNE :: Lens' DescribeBatchPredictions (Maybe Text)

The not equal to operator. The BatchPrediction results will have FilterVariable values not equal to the value specified with NE .

dbpNextToken :: Lens' DescribeBatchPredictions (Maybe Text)

An ID of the page in the paginated results.

dbpSortOrder :: Lens' DescribeBatchPredictions (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of MLModel s. * asc - Arranges the list in ascending order (A-Z, 0-9). * dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable .

dbpLimit :: Lens' DescribeBatchPredictions (Maybe Natural)

The number of pages of information to include in the result. The range of acceptable values is 1 through 100 . The default value is 100 .

dbpLT :: Lens' DescribeBatchPredictions (Maybe Text)

The less than operator. The BatchPrediction results will have FilterVariable values that are less than the value specified with LT .

dbpFilterVariable :: Lens' DescribeBatchPredictions (Maybe BatchPredictionFilterVariable)

Use one of the following variables to filter a list of BatchPrediction : * CreatedAt - Sets the search criteria to the BatchPrediction creation date. * Status - Sets the search criteria to the BatchPrediction status. * Name - Sets the search criteria to the contents of the BatchPrediction ____ Name . * IAMUser - Sets the search criteria to the user account that invoked the BatchPrediction creation. * MLModelId - Sets the search criteria to the MLModel used in the BatchPrediction . * DataSourceId - Sets the search criteria to the DataSource used in the BatchPrediction . * DataURI - Sets the search criteria to the data file(s) used in the BatchPrediction . The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

dbpLE :: Lens' DescribeBatchPredictions (Maybe Text)

The less than or equal to operator. The BatchPrediction results will have FilterVariable values that are less than or equal to the value specified with LE .

Destructuring the Response

describeBatchPredictionsResponse

Arguments

:: Int
dbpsrsResponseStatus

-> DescribeBatchPredictionsResponse

Creates a value of DescribeBatchPredictionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbpsrsResults - A list of BatchPrediction objects that meet the search criteria.

dbpsrsNextToken - The ID of the next page in the paginated results that indicates at least one more page follows.

dbpsrsResponseStatus - -- | The response status code.

data DescribeBatchPredictionsResponse

Represents the output of a DescribeBatchPredictions operation. The content is essentially a list of BatchPrediction s.

See: describeBatchPredictionsResponse smart constructor.

Response Lenses

dbpsrsResults :: Lens' DescribeBatchPredictionsResponse [BatchPrediction]

A list of BatchPrediction objects that meet the search criteria.

dbpsrsNextToken :: Lens' DescribeBatchPredictionsResponse (Maybe Text)

The ID of the next page in the paginated results that indicates at least one more page follows.

dbpsrsResponseStatus :: Lens' DescribeBatchPredictionsResponse Int

- | The response status code.