Network.AWS.MachineLearning.DescribeDataSources

Description

Returns a list of DataSource that match the search criteria in the request.

This operation returns paginated results.

Creating a Request

describeDataSources :: DescribeDataSources

Creates a value of DescribeDataSources with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddsEQ - The equal to operator. The DataSource results will have FilterVariable values that exactly match the value specified with EQ .

ddsGE - The greater than or equal to operator. The DataSource results will have FilterVariable values that are greater than or equal to the value specified with GE .

ddsPrefix - A string that is found at the beginning of a variable, such as Name or Id . For example, a DataSource could have the Name 2014-09-09-HolidayGiftMailer . To search for this DataSource , select Name for the FilterVariable and any of the following strings for the Prefix : * 2014-09 * 2014-09-09 * 2014-09-09-Holiday

ddsGT - The greater than operator. The DataSource results will have FilterVariable values that are greater than the value specified with GT .

ddsNE - The not equal to operator. The DataSource results will have FilterVariable values not equal to the value specified with NE .

ddsNextToken - The ID of the page in the paginated results.

ddsSortOrder - A two-value parameter that determines the sequence of the resulting list of DataSource . * asc - Arranges the list in ascending order (A-Z, 0-9). * dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable .

ddsLimit - The maximum number of DataSource to include in the result.

ddsLT - The less than operator. The DataSource results will have FilterVariable values that are less than the value specified with LT .

ddsFilterVariable - Use one of the following variables to filter a list of DataSource : * CreatedAt - Sets the search criteria to DataSource creation dates. * Status - Sets the search criteria to DataSource statuses. * Name - Sets the search criteria to the contents of DataSource ____ Name . * DataUri - Sets the search criteria to the URI of data files used to create the DataSource . The URI can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory. * IAMUser - Sets the search criteria to the user account that invoked the DataSource creation.

ddsLE - The less than or equal to operator. The DataSource results will have FilterVariable values that are less than or equal to the value specified with LE .

data DescribeDataSources

See: describeDataSources smart constructor.

Request Lenses

ddsEQ :: Lens' DescribeDataSources (Maybe Text)

The equal to operator. The DataSource results will have FilterVariable values that exactly match the value specified with EQ .

ddsGE :: Lens' DescribeDataSources (Maybe Text)

The greater than or equal to operator. The DataSource results will have FilterVariable values that are greater than or equal to the value specified with GE .

ddsPrefix :: Lens' DescribeDataSources (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id . For example, a DataSource could have the Name 2014-09-09-HolidayGiftMailer . To search for this DataSource , select Name for the FilterVariable and any of the following strings for the Prefix : * 2014-09 * 2014-09-09 * 2014-09-09-Holiday

ddsGT :: Lens' DescribeDataSources (Maybe Text)

The greater than operator. The DataSource results will have FilterVariable values that are greater than the value specified with GT .

ddsNE :: Lens' DescribeDataSources (Maybe Text)

The not equal to operator. The DataSource results will have FilterVariable values not equal to the value specified with NE .

ddsNextToken :: Lens' DescribeDataSources (Maybe Text)

The ID of the page in the paginated results.

ddsSortOrder :: Lens' DescribeDataSources (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of DataSource . * asc - Arranges the list in ascending order (A-Z, 0-9). * dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable .

ddsLimit :: Lens' DescribeDataSources (Maybe Natural)

The maximum number of DataSource to include in the result.

ddsLT :: Lens' DescribeDataSources (Maybe Text)

The less than operator. The DataSource results will have FilterVariable values that are less than the value specified with LT .

ddsFilterVariable :: Lens' DescribeDataSources (Maybe DataSourceFilterVariable)

Use one of the following variables to filter a list of DataSource : * CreatedAt - Sets the search criteria to DataSource creation dates. * Status - Sets the search criteria to DataSource statuses. * Name - Sets the search criteria to the contents of DataSource ____ Name . * DataUri - Sets the search criteria to the URI of data files used to create the DataSource . The URI can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory. * IAMUser - Sets the search criteria to the user account that invoked the DataSource creation.

ddsLE :: Lens' DescribeDataSources (Maybe Text)

The less than or equal to operator. The DataSource results will have FilterVariable values that are less than or equal to the value specified with LE .

Destructuring the Response

describeDataSourcesResponse

Arguments

:: Int
ddssrsResponseStatus

-> DescribeDataSourcesResponse

Creates a value of DescribeDataSourcesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddssrsResults - A list of DataSource that meet the search criteria.

ddssrsNextToken - An ID of the next page in the paginated results that indicates at least one more page follows.

ddssrsResponseStatus - -- | The response status code.

data DescribeDataSourcesResponse

Represents the query results from a DescribeDataSources operation. The content is essentially a list of DataSource .

See: describeDataSourcesResponse smart constructor.

Response Lenses

ddssrsResults :: Lens' DescribeDataSourcesResponse [DataSource]

A list of DataSource that meet the search criteria.

ddssrsNextToken :: Lens' DescribeDataSourcesResponse (Maybe Text)

An ID of the next page in the paginated results that indicates at least one more page follows.

ddssrsResponseStatus :: Lens' DescribeDataSourcesResponse Int

- | The response status code.