Network.AWS.MachineLearning.DescribeEvaluations

Description

Returns a list of DescribeEvaluations that match the search criteria in the request.

This operation returns paginated results.

Creating a Request

describeEvaluations :: DescribeEvaluations

Creates a value of DescribeEvaluations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deEQ - The equal to operator. The Evaluation results will have FilterVariable values that exactly match the value specified with EQ .

deGE - The greater than or equal to operator. The Evaluation results will have FilterVariable values that are greater than or equal to the value specified with GE .

dePrefix - A string that is found at the beginning of a variable, such as Name or Id . For example, an Evaluation could have the Name 2014-09-09-HolidayGiftMailer . To search for this Evaluation , select Name for the FilterVariable and any of the following strings for the Prefix : * 2014-09 * 2014-09-09 * 2014-09-09-Holiday

deGT - The greater than operator. The Evaluation results will have FilterVariable values that are greater than the value specified with GT .

deNE - The not equal to operator. The Evaluation results will have FilterVariable values not equal to the value specified with NE .

deNextToken - The ID of the page in the paginated results.

deSortOrder - A two-value parameter that determines the sequence of the resulting list of Evaluation . * asc - Arranges the list in ascending order (A-Z, 0-9). * dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable .

deLimit - The maximum number of Evaluation to include in the result.

deLT - The less than operator. The Evaluation results will have FilterVariable values that are less than the value specified with LT .

deFilterVariable - Use one of the following variable to filter a list of Evaluation objects: * CreatedAt - Sets the search criteria to the Evaluation creation date. * Status - Sets the search criteria to the Evaluation status. * Name - Sets the search criteria to the contents of Evaluation ____ Name . * IAMUser - Sets the search criteria to the user account that invoked an Evaluation . * MLModelId - Sets the search criteria to the MLModel that was evaluated. * DataSourceId - Sets the search criteria to the DataSource used in Evaluation . * DataUri - Sets the search criteria to the data file(s) used in Evaluation . The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

deLE - The less than or equal to operator. The Evaluation results will have FilterVariable values that are less than or equal to the value specified with LE .

data DescribeEvaluations

See: describeEvaluations smart constructor.

Request Lenses

deEQ :: Lens' DescribeEvaluations (Maybe Text)

The equal to operator. The Evaluation results will have FilterVariable values that exactly match the value specified with EQ .

deGE :: Lens' DescribeEvaluations (Maybe Text)

The greater than or equal to operator. The Evaluation results will have FilterVariable values that are greater than or equal to the value specified with GE .

dePrefix :: Lens' DescribeEvaluations (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id . For example, an Evaluation could have the Name 2014-09-09-HolidayGiftMailer . To search for this Evaluation , select Name for the FilterVariable and any of the following strings for the Prefix : * 2014-09 * 2014-09-09 * 2014-09-09-Holiday

deGT :: Lens' DescribeEvaluations (Maybe Text)

The greater than operator. The Evaluation results will have FilterVariable values that are greater than the value specified with GT .

deNE :: Lens' DescribeEvaluations (Maybe Text)

The not equal to operator. The Evaluation results will have FilterVariable values not equal to the value specified with NE .

deNextToken :: Lens' DescribeEvaluations (Maybe Text)

The ID of the page in the paginated results.

deSortOrder :: Lens' DescribeEvaluations (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of Evaluation . * asc - Arranges the list in ascending order (A-Z, 0-9). * dsc - Arranges the list in descending order (Z-A, 9-0). Results are sorted by FilterVariable .

deLimit :: Lens' DescribeEvaluations (Maybe Natural)

The maximum number of Evaluation to include in the result.

deLT :: Lens' DescribeEvaluations (Maybe Text)

The less than operator. The Evaluation results will have FilterVariable values that are less than the value specified with LT .

deFilterVariable :: Lens' DescribeEvaluations (Maybe EvaluationFilterVariable)

Use one of the following variable to filter a list of Evaluation objects: * CreatedAt - Sets the search criteria to the Evaluation creation date. * Status - Sets the search criteria to the Evaluation status. * Name - Sets the search criteria to the contents of Evaluation ____ Name . * IAMUser - Sets the search criteria to the user account that invoked an Evaluation . * MLModelId - Sets the search criteria to the MLModel that was evaluated. * DataSourceId - Sets the search criteria to the DataSource used in Evaluation . * DataUri - Sets the search criteria to the data file(s) used in Evaluation . The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

deLE :: Lens' DescribeEvaluations (Maybe Text)

The less than or equal to operator. The Evaluation results will have FilterVariable values that are less than or equal to the value specified with LE .

Destructuring the Response

describeEvaluationsResponse

Arguments

:: Int
desrsResponseStatus

-> DescribeEvaluationsResponse

Creates a value of DescribeEvaluationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsResults - A list of Evaluation that meet the search criteria.

desrsNextToken - The ID of the next page in the paginated results that indicates at least one more page follows.

desrsResponseStatus - -- | The response status code.

data DescribeEvaluationsResponse

Represents the query results from a DescribeEvaluations operation. The content is essentially a list of Evaluation .

See: describeEvaluationsResponse smart constructor.

Response Lenses

desrsResults :: Lens' DescribeEvaluationsResponse [Evaluation]

A list of Evaluation that meet the search criteria.

desrsNextToken :: Lens' DescribeEvaluationsResponse (Maybe Text)

The ID of the next page in the paginated results that indicates at least one more page follows.

desrsResponseStatus :: Lens' DescribeEvaluationsResponse Int

- | The response status code.