Network.AWS.MachineLearning.GetBatchPrediction

Description

Returns a BatchPrediction that includes detailed metadata, status, and data file information for a Batch Prediction request.

Creating a Request

getBatchPrediction

Arguments

:: Text
gbpBatchPredictionId

-> GetBatchPrediction

Creates a value of GetBatchPrediction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbpBatchPredictionId - An ID assigned to the BatchPrediction at creation.

data GetBatchPrediction

See: getBatchPrediction smart constructor.

Request Lenses

gbpBatchPredictionId :: Lens' GetBatchPrediction Text

An ID assigned to the BatchPrediction at creation.

Destructuring the Response

getBatchPredictionResponse

Arguments

:: Int
gbprsResponseStatus

-> GetBatchPredictionResponse

Creates a value of GetBatchPredictionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbprsStatus - The status of the BatchPrediction , which can be one of the following values: * PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate batch predictions. * INPROGRESS - The batch predictions are in progress. * FAILED - The request to perform a batch prediction did not run to completion. It is not usable. * COMPLETED - The batch prediction process completed successfully. * DELETED - The BatchPrediction is marked as deleted. It is not usable.

gbprsLastUpdatedAt - The time of the most recent edit to BatchPrediction . The time is expressed in epoch time.

gbprsCreatedAt - The time when the BatchPrediction was created. The time is expressed in epoch time.

gbprsComputeTime - The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the BatchPrediction , normalized and scaled on computation resources. ComputeTime is only available if the BatchPrediction is in the COMPLETED state.

gbprsInputDataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gbprsMLModelId - The ID of the MLModel that generated predictions for the BatchPrediction request.

gbprsBatchPredictionDataSourceId - The ID of the DataSource that was used to create the BatchPrediction .

gbprsTotalRecordCount - The number of total records that Amazon Machine Learning saw while processing the BatchPrediction .

gbprsStartedAt - The epoch time when Amazon Machine Learning marked the BatchPrediction as INPROGRESS . StartedAt isn't available if the BatchPrediction is in the PENDING state.

gbprsBatchPredictionId - An ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

gbprsFinishedAt - The epoch time when Amazon Machine Learning marked the BatchPrediction as COMPLETED or FAILED . FinishedAt is only available when the BatchPrediction is in the COMPLETED or FAILED state.

gbprsInvalidRecordCount - The number of invalid records that Amazon Machine Learning saw while processing the BatchPrediction .

gbprsCreatedByIAMUser - The AWS user account that invoked the BatchPrediction . The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gbprsName - A user-supplied name or description of the BatchPrediction .

gbprsLogURI - A link to the file that contains logs of the CreateBatchPrediction operation.

gbprsMessage - A description of the most recent details about processing the batch prediction request.

gbprsOutputURI - The location of an Amazon S3 bucket or directory to receive the operation results.

gbprsResponseStatus - -- | The response status code.

data GetBatchPredictionResponse

Represents the output of a GetBatchPrediction operation and describes a BatchPrediction .

See: getBatchPredictionResponse smart constructor.

Response Lenses

gbprsStatus :: Lens' GetBatchPredictionResponse (Maybe EntityStatus)

The status of the BatchPrediction , which can be one of the following values: * PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate batch predictions. * INPROGRESS - The batch predictions are in progress. * FAILED - The request to perform a batch prediction did not run to completion. It is not usable. * COMPLETED - The batch prediction process completed successfully. * DELETED - The BatchPrediction is marked as deleted. It is not usable.

gbprsLastUpdatedAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The time of the most recent edit to BatchPrediction . The time is expressed in epoch time.

gbprsCreatedAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The time when the BatchPrediction was created. The time is expressed in epoch time.

gbprsComputeTime :: Lens' GetBatchPredictionResponse (Maybe Integer)

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the BatchPrediction , normalized and scaled on computation resources. ComputeTime is only available if the BatchPrediction is in the COMPLETED state.

gbprsInputDataLocationS3 :: Lens' GetBatchPredictionResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gbprsMLModelId :: Lens' GetBatchPredictionResponse (Maybe Text)

The ID of the MLModel that generated predictions for the BatchPrediction request.

gbprsBatchPredictionDataSourceId :: Lens' GetBatchPredictionResponse (Maybe Text)

The ID of the DataSource that was used to create the BatchPrediction .

gbprsTotalRecordCount :: Lens' GetBatchPredictionResponse (Maybe Integer)

The number of total records that Amazon Machine Learning saw while processing the BatchPrediction .

gbprsStartedAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the BatchPrediction as INPROGRESS . StartedAt isn't available if the BatchPrediction is in the PENDING state.

gbprsBatchPredictionId :: Lens' GetBatchPredictionResponse (Maybe Text)

An ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

gbprsFinishedAt :: Lens' GetBatchPredictionResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the BatchPrediction as COMPLETED or FAILED . FinishedAt is only available when the BatchPrediction is in the COMPLETED or FAILED state.

gbprsInvalidRecordCount :: Lens' GetBatchPredictionResponse (Maybe Integer)

The number of invalid records that Amazon Machine Learning saw while processing the BatchPrediction .

gbprsCreatedByIAMUser :: Lens' GetBatchPredictionResponse (Maybe Text)

The AWS user account that invoked the BatchPrediction . The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gbprsName :: Lens' GetBatchPredictionResponse (Maybe Text)

A user-supplied name or description of the BatchPrediction .

gbprsLogURI :: Lens' GetBatchPredictionResponse (Maybe Text)

A link to the file that contains logs of the CreateBatchPrediction operation.

gbprsMessage :: Lens' GetBatchPredictionResponse (Maybe Text)

A description of the most recent details about processing the batch prediction request.

gbprsOutputURI :: Lens' GetBatchPredictionResponse (Maybe Text)

The location of an Amazon S3 bucket or directory to receive the operation results.

gbprsResponseStatus :: Lens' GetBatchPredictionResponse Int

- | The response status code.