Network.AWS.MachineLearning.GetDataSource

Description

Returns a DataSource that includes metadata and data file information, as well as the current status of the DataSource .

GetDataSource provides results in normal or verbose format. The verbose format adds the schema description and the list of files pointed to by the DataSource to the normal format.

Creating a Request

getDataSource

Arguments

:: Text
gdsDataSourceId

-> GetDataSource

Creates a value of GetDataSource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsVerbose - Specifies whether the GetDataSource operation should return DataSourceSchema . If true, DataSourceSchema is returned. If false, DataSourceSchema is not returned.

gdsDataSourceId - The ID assigned to the DataSource at creation.

data GetDataSource

See: getDataSource smart constructor.

Request Lenses

gdsVerbose :: Lens' GetDataSource (Maybe Bool)

Specifies whether the GetDataSource operation should return DataSourceSchema . If true, DataSourceSchema is returned. If false, DataSourceSchema is not returned.

gdsDataSourceId :: Lens' GetDataSource Text

The ID assigned to the DataSource at creation.

Destructuring the Response

getDataSourceResponse

Arguments

:: Int
gdsrsResponseStatus

-> GetDataSourceResponse

Creates a value of GetDataSourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsrsStatus - The current status of the DataSource . This element can have one of the following values: * PENDING - Amazon ML submitted a request to create a DataSource . * INPROGRESS - The creation process is underway. * FAILED - The request to create a DataSource did not run to completion. It is not usable. * COMPLETED - The creation process completed successfully. * DELETED - The DataSource is marked as deleted. It is not usable.

gdsrsNumberOfFiles - The number of data files referenced by the DataSource .

gdsrsLastUpdatedAt - The time of the most recent edit to the DataSource . The time is expressed in epoch time.

gdsrsCreatedAt - The time that the DataSource was created. The time is expressed in epoch time.

gdsrsComputeTime - The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the DataSource , normalized and scaled on computation resources. ComputeTime is only available if the DataSource is in the COMPLETED state and the ComputeStatistics is set to true.

gdsrsDataSourceId - The ID assigned to the DataSource at creation. This value should be identical to the value of the DataSourceId in the request.

gdsrsRDSMetadata - Undocumented member.

gdsrsDataSizeInBytes - The total size of observations in the data files.

gdsrsDataSourceSchema - The schema used by all of the data files of this DataSource .

gdsrsStartedAt - The epoch time when Amazon Machine Learning marked the DataSource as INPROGRESS . StartedAt isn't available if the DataSource is in the PENDING state.

gdsrsFinishedAt - The epoch time when Amazon Machine Learning marked the DataSource as COMPLETED or FAILED . FinishedAt is only available when the DataSource is in the COMPLETED or FAILED state.

gdsrsCreatedByIAMUser - The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gdsrsName - A user-supplied name or description of the DataSource .

gdsrsLogURI - A link to the file containing logs of CreateDataSourceFrom* operations.

gdsrsDataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gdsrsComputeStatistics - The parameter is true if statistics need to be generated from the observation data.

gdsrsMessage - The user-supplied description of the most recent details about creating the DataSource .

gdsrsRedshiftMetadata - Undocumented member.

gdsrsDataRearrangement - A JSON string that represents the splitting and rearrangement requirement used when this DataSource was created.

gdsrsRoleARN - Undocumented member.

gdsrsResponseStatus - -- | The response status code.

data GetDataSourceResponse

Represents the output of a GetDataSource operation and describes a DataSource .

See: getDataSourceResponse smart constructor.

Response Lenses

gdsrsStatus :: Lens' GetDataSourceResponse (Maybe EntityStatus)

The current status of the DataSource . This element can have one of the following values: * PENDING - Amazon ML submitted a request to create a DataSource . * INPROGRESS - The creation process is underway. * FAILED - The request to create a DataSource did not run to completion. It is not usable. * COMPLETED - The creation process completed successfully. * DELETED - The DataSource is marked as deleted. It is not usable.

gdsrsNumberOfFiles :: Lens' GetDataSourceResponse (Maybe Integer)

The number of data files referenced by the DataSource .

gdsrsLastUpdatedAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The time of the most recent edit to the DataSource . The time is expressed in epoch time.

gdsrsCreatedAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The time that the DataSource was created. The time is expressed in epoch time.

gdsrsComputeTime :: Lens' GetDataSourceResponse (Maybe Integer)

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the DataSource , normalized and scaled on computation resources. ComputeTime is only available if the DataSource is in the COMPLETED state and the ComputeStatistics is set to true.

gdsrsDataSourceId :: Lens' GetDataSourceResponse (Maybe Text)

The ID assigned to the DataSource at creation. This value should be identical to the value of the DataSourceId in the request.

gdsrsRDSMetadata :: Lens' GetDataSourceResponse (Maybe RDSMetadata)

Undocumented member.

gdsrsDataSizeInBytes :: Lens' GetDataSourceResponse (Maybe Integer)

The total size of observations in the data files.

gdsrsDataSourceSchema :: Lens' GetDataSourceResponse (Maybe Text)

The schema used by all of the data files of this DataSource .

gdsrsStartedAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the DataSource as INPROGRESS . StartedAt isn't available if the DataSource is in the PENDING state.

gdsrsFinishedAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the DataSource as COMPLETED or FAILED . FinishedAt is only available when the DataSource is in the COMPLETED or FAILED state.

gdsrsCreatedByIAMUser :: Lens' GetDataSourceResponse (Maybe Text)

The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gdsrsName :: Lens' GetDataSourceResponse (Maybe Text)

A user-supplied name or description of the DataSource .

gdsrsLogURI :: Lens' GetDataSourceResponse (Maybe Text)

A link to the file containing logs of CreateDataSourceFrom* operations.

gdsrsDataLocationS3 :: Lens' GetDataSourceResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gdsrsComputeStatistics :: Lens' GetDataSourceResponse (Maybe Bool)

The parameter is true if statistics need to be generated from the observation data.

gdsrsMessage :: Lens' GetDataSourceResponse (Maybe Text)

The user-supplied description of the most recent details about creating the DataSource .

gdsrsRedshiftMetadata :: Lens' GetDataSourceResponse (Maybe RedshiftMetadata)

Undocumented member.

gdsrsDataRearrangement :: Lens' GetDataSourceResponse (Maybe Text)

A JSON string that represents the splitting and rearrangement requirement used when this DataSource was created.

gdsrsRoleARN :: Lens' GetDataSourceResponse (Maybe Text)

Undocumented member.

gdsrsResponseStatus :: Lens' GetDataSourceResponse Int

- | The response status code.