Network.AWS.MachineLearning.GetEvaluation

Description

Returns an Evaluation that includes metadata as well as the current status of the Evaluation .

Creating a Request

getEvaluation

Arguments

:: Text
geEvaluationId

-> GetEvaluation

Creates a value of GetEvaluation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

geEvaluationId - The ID of the Evaluation to retrieve. The evaluation of each MLModel is recorded and cataloged. The ID provides the means to access the information.

data GetEvaluation

See: getEvaluation smart constructor.

Request Lenses

geEvaluationId :: Lens' GetEvaluation Text

The ID of the Evaluation to retrieve. The evaluation of each MLModel is recorded and cataloged. The ID provides the means to access the information.

Destructuring the Response

getEvaluationResponse

Arguments

:: Int
gersResponseStatus

-> GetEvaluationResponse

Creates a value of GetEvaluationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gersStatus - The status of the evaluation. This element can have one of the following values: * PENDING - Amazon Machine Language (Amazon ML) submitted a request to evaluate an MLModel . * INPROGRESS - The evaluation is underway. * FAILED - The request to evaluate an MLModel did not run to completion. It is not usable. * COMPLETED - The evaluation process completed successfully. * DELETED - The Evaluation is marked as deleted. It is not usable.

gersPerformanceMetrics - Measurements of how well the MLModel performed using observations referenced by the DataSource . One of the following metric is returned based on the type of the MLModel : * BinaryAUC: A binary MLModel uses the Area Under the Curve (AUC) technique to measure performance. * RegressionRMSE: A regression MLModel uses the Root Mean Square Error (RMSE) technique to measure performance. RMSE measures the difference between predicted and actual values for a single variable. * MulticlassAvgFScore: A multiclass MLModel uses the F1 score technique to measure performance. For more information about performance metrics, please see the Amazon Machine Learning Developer Guide .

gersLastUpdatedAt - The time of the most recent edit to the Evaluation . The time is expressed in epoch time.

gersCreatedAt - The time that the Evaluation was created. The time is expressed in epoch time.

gersComputeTime - The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the Evaluation , normalized and scaled on computation resources. ComputeTime is only available if the Evaluation is in the COMPLETED state.

gersInputDataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gersMLModelId - The ID of the MLModel that was the focus of the evaluation.

gersStartedAt - The epoch time when Amazon Machine Learning marked the Evaluation as INPROGRESS . StartedAt isn't available if the Evaluation is in the PENDING state.

gersFinishedAt - The epoch time when Amazon Machine Learning marked the Evaluation as COMPLETED or FAILED . FinishedAt is only available when the Evaluation is in the COMPLETED or FAILED state.

gersCreatedByIAMUser - The AWS user account that invoked the evaluation. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gersName - A user-supplied name or description of the Evaluation .

gersLogURI - A link to the file that contains logs of the CreateEvaluation operation.

gersEvaluationId - The evaluation ID which is same as the EvaluationId in the request.

gersMessage - A description of the most recent details about evaluating the MLModel .

gersEvaluationDataSourceId - The DataSource used for this evaluation.

gersResponseStatus - -- | The response status code.

data GetEvaluationResponse

Represents the output of a GetEvaluation operation and describes an Evaluation .

See: getEvaluationResponse smart constructor.

Response Lenses

gersStatus :: Lens' GetEvaluationResponse (Maybe EntityStatus)

The status of the evaluation. This element can have one of the following values: * PENDING - Amazon Machine Language (Amazon ML) submitted a request to evaluate an MLModel . * INPROGRESS - The evaluation is underway. * FAILED - The request to evaluate an MLModel did not run to completion. It is not usable. * COMPLETED - The evaluation process completed successfully. * DELETED - The Evaluation is marked as deleted. It is not usable.

gersPerformanceMetrics :: Lens' GetEvaluationResponse (Maybe PerformanceMetrics)

Measurements of how well the MLModel performed using observations referenced by the DataSource . One of the following metric is returned based on the type of the MLModel : * BinaryAUC: A binary MLModel uses the Area Under the Curve (AUC) technique to measure performance. * RegressionRMSE: A regression MLModel uses the Root Mean Square Error (RMSE) technique to measure performance. RMSE measures the difference between predicted and actual values for a single variable. * MulticlassAvgFScore: A multiclass MLModel uses the F1 score technique to measure performance. For more information about performance metrics, please see the Amazon Machine Learning Developer Guide .

gersLastUpdatedAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The time of the most recent edit to the Evaluation . The time is expressed in epoch time.

gersCreatedAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The time that the Evaluation was created. The time is expressed in epoch time.

gersComputeTime :: Lens' GetEvaluationResponse (Maybe Integer)

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the Evaluation , normalized and scaled on computation resources. ComputeTime is only available if the Evaluation is in the COMPLETED state.

gersInputDataLocationS3 :: Lens' GetEvaluationResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gersMLModelId :: Lens' GetEvaluationResponse (Maybe Text)

The ID of the MLModel that was the focus of the evaluation.

gersStartedAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the Evaluation as INPROGRESS . StartedAt isn't available if the Evaluation is in the PENDING state.

gersFinishedAt :: Lens' GetEvaluationResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the Evaluation as COMPLETED or FAILED . FinishedAt is only available when the Evaluation is in the COMPLETED or FAILED state.

gersCreatedByIAMUser :: Lens' GetEvaluationResponse (Maybe Text)

The AWS user account that invoked the evaluation. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gersName :: Lens' GetEvaluationResponse (Maybe Text)

A user-supplied name or description of the Evaluation .

gersLogURI :: Lens' GetEvaluationResponse (Maybe Text)

A link to the file that contains logs of the CreateEvaluation operation.

gersEvaluationId :: Lens' GetEvaluationResponse (Maybe Text)

The evaluation ID which is same as the EvaluationId in the request.

gersMessage :: Lens' GetEvaluationResponse (Maybe Text)

A description of the most recent details about evaluating the MLModel .

gersEvaluationDataSourceId :: Lens' GetEvaluationResponse (Maybe Text)

The DataSource used for this evaluation.

gersResponseStatus :: Lens' GetEvaluationResponse Int

- | The response status code.