Network.AWS.MachineLearning.GetMLModel

Description

Returns an MLModel that includes detailed metadata, data source information, and the current status of the MLModel .

GetMLModel provides results in normal or verbose format.

Creating a Request

getMLModel

Arguments

:: Text
gmlmMLModelId

-> GetMLModel

Creates a value of GetMLModel with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmlmVerbose - Specifies whether the GetMLModel operation should return Recipe . If true, Recipe is returned. If false, Recipe is not returned.

gmlmMLModelId - The ID assigned to the MLModel at creation.

data GetMLModel

See: getMLModel smart constructor.

Request Lenses

gmlmVerbose :: Lens' GetMLModel (Maybe Bool)

Specifies whether the GetMLModel operation should return Recipe . If true, Recipe is returned. If false, Recipe is not returned.

gmlmMLModelId :: Lens' GetMLModel Text

The ID assigned to the MLModel at creation.

Destructuring the Response

getMLModelResponse

Arguments

:: Int
gmlmrsResponseStatus

-> GetMLModelResponse

Creates a value of GetMLModelResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gmlmrsStatus - The current status of the MLModel . This element can have one of the following values: * PENDING - Amazon Machine Learning (Amazon ML) submitted a request to describe a MLModel . * INPROGRESS - The request is processing. * FAILED - The request did not run to completion. The ML model isn't usable. * COMPLETED - The request completed successfully. * DELETED - The MLModel is marked as deleted. It isn't usable.

gmlmrsLastUpdatedAt - The time of the most recent edit to the MLModel . The time is expressed in epoch time.

gmlmrsTrainingParameters - A list of the training parameters in the MLModel . The list is implemented as a map of key-value pairs. The following is the current set of training parameters: * sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance. The value is an integer that ranges from 100000 to 2147483648 . The default value is 33554432 . * sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel . The value is an integer that ranges from 1 to 10000 . The default value is 10 . * sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none . The default value is none . We strongly recommend that you shuffle your data. * sgd.l1RegularizationAmount - The coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08 . The value is a double that ranges from 0 to MAX_DOUBLE . The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly. * sgd.l2RegularizationAmount - The coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08 . The value is a double that ranges from 0 to MAX_DOUBLE . The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

gmlmrsScoreThresholdLastUpdatedAt - The time of the most recent edit to the ScoreThreshold . The time is expressed in epoch time.

gmlmrsCreatedAt - The time that the MLModel was created. The time is expressed in epoch time.

gmlmrsComputeTime - The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the MLModel , normalized and scaled on computation resources. ComputeTime is only available if the MLModel is in the COMPLETED state.

gmlmrsRecipe - The recipe to use when training the MLModel . The Recipe provides detailed information about the observation data to use during training, and manipulations to perform on the observation data during training.

gmlmrsInputDataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gmlmrsMLModelId - The MLModel ID, which is same as the MLModelId in the request.

gmlmrsSizeInBytes - Undocumented member.

gmlmrsSchema - The schema used by all of the data files referenced by the DataSource .

gmlmrsStartedAt - The epoch time when Amazon Machine Learning marked the MLModel as INPROGRESS . StartedAt isn't available if the MLModel is in the PENDING state.

gmlmrsScoreThreshold - The scoring threshold is used in binary classification MLModel models. It marks the boundary between a positive prediction and a negative prediction. Output values greater than or equal to the threshold receive a positive result from the MLModel, such as true . Output values less than the threshold receive a negative response from the MLModel, such as false .

gmlmrsFinishedAt - The epoch time when Amazon Machine Learning marked the MLModel as COMPLETED or FAILED . FinishedAt is only available when the MLModel is in the COMPLETED or FAILED state.

gmlmrsCreatedByIAMUser - The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gmlmrsName - A user-supplied name or description of the MLModel .

gmlmrsLogURI - A link to the file that contains logs of the CreateMLModel operation.

gmlmrsEndpointInfo - The current endpoint of the MLModel

gmlmrsTrainingDataSourceId - The ID of the training DataSource .

gmlmrsMessage - A description of the most recent details about accessing the MLModel .

gmlmrsMLModelType - Identifies the MLModel category. The following are the available types: * REGRESSION -- Produces a numeric result. For example, "What price should a house be listed at?" * BINARY -- Produces one of two possible results. For example, "Is this an e-commerce website?" * MULTICLASS -- Produces one of several possible results. For example, "Is this a HIGH, LOW or MEDIUM risk trade?"

gmlmrsResponseStatus - -- | The response status code.

data GetMLModelResponse

Represents the output of a GetMLModel operation, and provides detailed information about a MLModel .

See: getMLModelResponse smart constructor.

Response Lenses

gmlmrsStatus :: Lens' GetMLModelResponse (Maybe EntityStatus)

The current status of the MLModel . This element can have one of the following values: * PENDING - Amazon Machine Learning (Amazon ML) submitted a request to describe a MLModel . * INPROGRESS - The request is processing. * FAILED - The request did not run to completion. The ML model isn't usable. * COMPLETED - The request completed successfully. * DELETED - The MLModel is marked as deleted. It isn't usable.

gmlmrsLastUpdatedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The time of the most recent edit to the MLModel . The time is expressed in epoch time.

gmlmrsTrainingParameters :: Lens' GetMLModelResponse (HashMap Text Text)

A list of the training parameters in the MLModel . The list is implemented as a map of key-value pairs. The following is the current set of training parameters: * sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance. The value is an integer that ranges from 100000 to 2147483648 . The default value is 33554432 . * sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel . The value is an integer that ranges from 1 to 10000 . The default value is 10 . * sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none . The default value is none . We strongly recommend that you shuffle your data. * sgd.l1RegularizationAmount - The coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08 . The value is a double that ranges from 0 to MAX_DOUBLE . The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly. * sgd.l2RegularizationAmount - The coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08 . The value is a double that ranges from 0 to MAX_DOUBLE . The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

gmlmrsScoreThresholdLastUpdatedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The time of the most recent edit to the ScoreThreshold . The time is expressed in epoch time.

gmlmrsCreatedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The time that the MLModel was created. The time is expressed in epoch time.

gmlmrsComputeTime :: Lens' GetMLModelResponse (Maybe Integer)

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the MLModel , normalized and scaled on computation resources. ComputeTime is only available if the MLModel is in the COMPLETED state.

gmlmrsRecipe :: Lens' GetMLModelResponse (Maybe Text)

The recipe to use when training the MLModel . The Recipe provides detailed information about the observation data to use during training, and manipulations to perform on the observation data during training.

gmlmrsInputDataLocationS3 :: Lens' GetMLModelResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

gmlmrsMLModelId :: Lens' GetMLModelResponse (Maybe Text)

The MLModel ID, which is same as the MLModelId in the request.

gmlmrsSizeInBytes :: Lens' GetMLModelResponse (Maybe Integer)

Undocumented member.

gmlmrsSchema :: Lens' GetMLModelResponse (Maybe Text)

The schema used by all of the data files referenced by the DataSource .

gmlmrsStartedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the MLModel as INPROGRESS . StartedAt isn't available if the MLModel is in the PENDING state.

gmlmrsScoreThreshold :: Lens' GetMLModelResponse (Maybe Double)

The scoring threshold is used in binary classification MLModel models. It marks the boundary between a positive prediction and a negative prediction. Output values greater than or equal to the threshold receive a positive result from the MLModel, such as true . Output values less than the threshold receive a negative response from the MLModel, such as false .

gmlmrsFinishedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the MLModel as COMPLETED or FAILED . FinishedAt is only available when the MLModel is in the COMPLETED or FAILED state.

gmlmrsCreatedByIAMUser :: Lens' GetMLModelResponse (Maybe Text)

The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

gmlmrsName :: Lens' GetMLModelResponse (Maybe Text)

A user-supplied name or description of the MLModel .

gmlmrsLogURI :: Lens' GetMLModelResponse (Maybe Text)

A link to the file that contains logs of the CreateMLModel operation.

gmlmrsEndpointInfo :: Lens' GetMLModelResponse (Maybe RealtimeEndpointInfo)

The current endpoint of the MLModel

gmlmrsTrainingDataSourceId :: Lens' GetMLModelResponse (Maybe Text)

The ID of the training DataSource .

gmlmrsMessage :: Lens' GetMLModelResponse (Maybe Text)

A description of the most recent details about accessing the MLModel .

gmlmrsMLModelType :: Lens' GetMLModelResponse (Maybe MLModelType)

Identifies the MLModel category. The following are the available types: * REGRESSION -- Produces a numeric result. For example, "What price should a house be listed at?" * BINARY -- Produces one of two possible results. For example, "Is this an e-commerce website?" * MULTICLASS -- Produces one of several possible results. For example, "Is this a HIGH, LOW or MEDIUM risk trade?"

gmlmrsResponseStatus :: Lens' GetMLModelResponse Int

- | The response status code.