Network.AWS.MachineLearning.UpdateBatchPrediction

Description

Updates the BatchPredictionName of a BatchPrediction .

You can use the GetBatchPrediction operation to view the contents of the updated data element.

Creating a Request

updateBatchPrediction

Arguments

:: Text
ubpBatchPredictionId

-> Text
ubpBatchPredictionName

-> UpdateBatchPrediction

Creates a value of UpdateBatchPrediction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ubpBatchPredictionId - The ID assigned to the BatchPrediction during creation.

ubpBatchPredictionName - A new user-supplied name or description of the BatchPrediction .

data UpdateBatchPrediction

See: updateBatchPrediction smart constructor.

Request Lenses

ubpBatchPredictionId :: Lens' UpdateBatchPrediction Text

The ID assigned to the BatchPrediction during creation.

ubpBatchPredictionName :: Lens' UpdateBatchPrediction Text

A new user-supplied name or description of the BatchPrediction .

Destructuring the Response

updateBatchPredictionResponse

Arguments

:: Int
ubprsResponseStatus

-> UpdateBatchPredictionResponse

Creates a value of UpdateBatchPredictionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ubprsBatchPredictionId - The ID assigned to the BatchPrediction during creation. This value should be identical to the value of the BatchPredictionId in the request.

ubprsResponseStatus - -- | The response status code.

data UpdateBatchPredictionResponse

Represents the output of an UpdateBatchPrediction operation.

You can see the updated content by using the GetBatchPrediction operation.

See: updateBatchPredictionResponse smart constructor.

Response Lenses

ubprsBatchPredictionId :: Lens' UpdateBatchPredictionResponse (Maybe Text)

The ID assigned to the BatchPrediction during creation. This value should be identical to the value of the BatchPredictionId in the request.

ubprsResponseStatus :: Lens' UpdateBatchPredictionResponse Int

- | The response status code.