Network.AWS.MachineLearning.UpdateEvaluation

Description

Updates the EvaluationName of an Evaluation .

You can use the GetEvaluation operation to view the contents of the updated data element.

Creating a Request

updateEvaluation

Arguments

:: Text
ueEvaluationId

-> Text
ueEvaluationName

-> UpdateEvaluation

Creates a value of UpdateEvaluation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ueEvaluationId - The ID assigned to the Evaluation during creation.

ueEvaluationName - A new user-supplied name or description of the Evaluation that will replace the current content.

data UpdateEvaluation

See: updateEvaluation smart constructor.

Request Lenses

ueEvaluationId :: Lens' UpdateEvaluation Text

The ID assigned to the Evaluation during creation.

ueEvaluationName :: Lens' UpdateEvaluation Text

A new user-supplied name or description of the Evaluation that will replace the current content.

Destructuring the Response

updateEvaluationResponse

Arguments

:: Int
uersResponseStatus

-> UpdateEvaluationResponse

Creates a value of UpdateEvaluationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uersEvaluationId - The ID assigned to the Evaluation during creation. This value should be identical to the value of the Evaluation in the request.

uersResponseStatus - -- | The response status code.

data UpdateEvaluationResponse

Represents the output of an UpdateEvaluation operation.

You can see the updated content by using the GetEvaluation operation.

See: updateEvaluationResponse smart constructor.

Response Lenses

uersEvaluationId :: Lens' UpdateEvaluationResponse (Maybe Text)

The ID assigned to the Evaluation during creation. This value should be identical to the value of the Evaluation in the request.

uersResponseStatus :: Lens' UpdateEvaluationResponse Int

- | The response status code.